refactor(test): split fsrs_test.go monolith into co-located *_test.go files
Pure file reorganization — no test added, removed, or modified.
fsrs_test.go (4999 lines) split into 14 co-located *_test.go files
(avg ~335 lines, none exceeding 1000). All 87 functions moved
byte-identical; package fsrs (white-box) preserved.

go vet clean, go test 91 PASS / 0 FAIL (identical to baseline).
