+++ fsrs_test.go
@@ -1,17 +1,11 @@
 package fsrs
 
 import (
-	"errors"
 	"math"
 	"reflect"
 	"testing"
 	"time"
 )
-
-func roundFloat(val float64, precision uint) float64 {
-	ratio := math.Pow(10, float64(precision))
-	return math.Round(val*ratio) / ratio
-}
 
 func TestBasicSchedulerExample(t *testing.T) {
 	p := DefaultParam()
@@ -207,60 +201,6 @@ func TestNextStatesMemoryStateLongTerm(t *testing.T) {
 	}
 }
 
-func TestNextInterval(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	var intervalList []float64
-	for i := 1; i <= 10; i++ {
-		fsrs.RequestRetention = float64(i) / 10
-		intervalList = append(intervalList, fsrs.nextInterval(1, 0))
-	}
-	wantIntervalList := []float64{36500, 34793, 2508, 387, 90, 27, 9, 3, 1, 1}
-	if !reflect.DeepEqual(intervalList, wantIntervalList) {
-		t.Errorf("expected:%v, got:%v", wantIntervalList, intervalList)
-	}
-}
-
-func TestNextIntervalRaw(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-
-	interval := fsrs.nextIntervalRaw(0.212)
-	if interval >= 1.0 {
-		t.Errorf("nextIntervalRaw(0.212) should return < 1 day, got=%v", interval)
-	}
-	if interval <= 0 {
-		t.Errorf("nextIntervalRaw(0.212) should return > 0, got=%v", interval)
-	}
-
-	interval = fsrs.nextIntervalRaw(1.0)
-	if interval < 0.99 {
-		t.Errorf("nextIntervalRaw(1.0) should return ~1 day, got=%v", interval)
-	}
-
-	interval = fsrs.nextIntervalRaw(0.001)
-	if interval >= 0.5 {
-		t.Errorf("nextIntervalRaw(0.001) should return < 0.5 days, got=%v", interval)
-	}
-
-	interval = fsrs.nextIntervalRaw(sMin)
-	if interval <= 0 || math.IsNaN(interval) || math.IsInf(interval, 0) {
-		t.Errorf("nextIntervalRaw(sMin) should return finite positive, got=%v", interval)
-	}
-
-	interval = fsrs.nextIntervalRaw(sMax)
-	if interval <= 0 || math.IsNaN(interval) || math.IsInf(interval, 0) {
-		t.Errorf("nextIntervalRaw(sMax) should return finite positive, got=%v", interval)
-	}
-
-	decay, factor := p.decayAndFactor()
-	boundaryStab := 0.5 * factor / (math.Pow(p.RequestRetention, 1/decay) - 1)
-	interval = fsrs.nextIntervalRaw(boundaryStab)
-	if math.Abs(interval-0.5) > 1e-9 {
-		t.Errorf("nextIntervalRaw at boundary stability should ≈ 0.5, got=%v", interval)
-	}
-}
-
 func TestLongTermScheduler(t *testing.T) {
 	p := DefaultParam()
 	p.EnableShortTerm = false
@@ -305,672 +245,6 @@ func TestLongTermScheduler(t *testing.T) {
 	}
 }
 
-func TestDateDiffInDays(t *testing.T) {
-	tests := []struct {
-		name     string
-		last     time.Time
-		now      time.Time
-		expected uint64
-	}{
-		{"same day different hours", time.Date(2032, 1, 15, 12, 30, 0, 0, time.UTC), time.Date(2032, 1, 15, 23, 59, 0, 0, time.UTC), 0},
-		{"next day", time.Date(2032, 1, 15, 0, 0, 0, 0, time.UTC), time.Date(2032, 1, 16, 0, 0, 0, 0, time.UTC), 1},
-		{"end of month", time.Date(2032, 1, 31, 12, 0, 0, 0, time.UTC), time.Date(2032, 2, 1, 0, 0, 0, 0, time.UTC), 1},
-		{"year boundary", time.Date(2032, 12, 31, 18, 0, 0, 0, time.UTC), time.Date(2033, 1, 1, 0, 0, 0, 0, time.UTC), 1},
-		{"large span", time.Date(2032, 1, 1, 0, 0, 0, 0, time.UTC), time.Date(2033, 1, 1, 0, 0, 0, 0, time.UTC), 366},
-		{"CET midnight crosses UTC day", time.Date(2032, 1, 15, 0, 30, 0, 0, time.FixedZone("CET", 3600)), time.Date(2032, 1, 15, 12, 0, 0, 0, time.FixedZone("CET", 3600)), 1},
-		{"CET evening to next UTC day", time.Date(2032, 1, 14, 23, 30, 0, 0, time.FixedZone("CET", 3600)), time.Date(2032, 1, 15, 1, 0, 0, 0, time.FixedZone("CET", 3600)), 1},
-		{"different local days but UTC boundary matters", time.Date(2032, 1, 15, 23, 0, 0, 0, time.UTC), time.Date(2032, 1, 16, 1, 0, 0, 0, time.FixedZone("EST", -18000)), 1},
-		{"multiple days", time.Date(2032, 1, 10, 0, 0, 0, 0, time.UTC), time.Date(2032, 1, 20, 0, 0, 0, 0, time.UTC), 10},
-		{"reversed dates returns 0", time.Date(2032, 1, 20, 0, 0, 0, 0, time.UTC), time.Date(2032, 1, 10, 0, 0, 0, 0, time.UTC), 0},
-		{"reversed same day returns 0", time.Date(2032, 1, 15, 18, 0, 0, 0, time.UTC), time.Date(2032, 1, 15, 6, 0, 0, 0, time.UTC), 0},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := dateDiffInDays(tt.last, tt.now)
-			if got != tt.expected {
-				t.Errorf("dateDiffInDays(%v, %v) = %d, want %d", tt.last, tt.now, got, tt.expected)
-			}
-		})
-	}
-}
-
-func TestNewCard(t *testing.T) {
-	t.Run("default sets Due to current time", func(t *testing.T) {
-		before := time.Now()
-		card := NewCard()
-		after := time.Now()
-		if card.Due.Before(before) || card.Due.After(after) {
-			t.Errorf("expected Due between %v and %v, got=%v", before, after, card.Due)
-		}
-	})
-
-	t.Run("explicit time sets Due to provided value", func(t *testing.T) {
-		specificTime := time.Date(2024, 6, 15, 10, 30, 0, 0, time.UTC)
-		card := NewCard(specificTime)
-		if !card.Due.Equal(specificTime) {
-			t.Errorf("expected Due=%v, got=%v", specificTime, card.Due)
-		}
-		if card.Stability != 0 || card.Difficulty != 0 {
-			t.Errorf("expected zero Stability/Difficulty, got s=%v d=%v", card.Stability, card.Difficulty)
-		}
-		if card.State != 0 || card.Reps != 0 || card.Lapses != 0 {
-			t.Errorf("expected zero State/Reps/Lapses, got state=%v reps=%d lapses=%d", card.State, card.Reps, card.Lapses)
-		}
-	})
-
-	t.Run("multiple variadic args uses only first", func(t *testing.T) {
-		first := time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		second := time.Date(2025, 6, 15, 0, 0, 0, 0, time.UTC)
-		card := NewCard(first, second)
-		if !card.Due.Equal(first) {
-			t.Errorf("expected Due=%v (first arg), got=%v", first, card.Due)
-		}
-	})
-
-	t.Run("zero time sets Due to zero", func(t *testing.T) {
-		card := NewCard(time.Time{})
-		if !card.Due.IsZero() {
-			t.Errorf("expected zero Due, got=%v", card.Due)
-		}
-	})
-
-	t.Run("other fields remain zero", func(t *testing.T) {
-		card := NewCard()
-		if card.Stability != 0 || card.Difficulty != 0 {
-			t.Errorf("expected zero Stability/Difficulty, got s=%v d=%v", card.Stability, card.Difficulty)
-		}
-		if card.State != 0 || card.Reps != 0 || card.Lapses != 0 {
-			t.Errorf("expected zero State/Reps/Lapses, got state=%v reps=%d lapses=%d", card.State, card.Reps, card.Lapses)
-		}
-	})
-}
-
-func TestClipParameters(t *testing.T) {
-	t.Run("default weights unchanged after clipping", func(t *testing.T) {
-		p := DefaultParam()
-		original := p.W
-		clipParameters(&p)
-		for i := range p.W {
-			if p.W[i] != original[i] {
-				t.Errorf("W[%d]: clipping changed default weight from %v to %v", i, original[i], p.W[i])
-			}
-		}
-	})
-
-	t.Run("extreme weights are clipped to range", func(t *testing.T) {
-		p := DefaultParam()
-		p.W[0] = -999
-		p.W[1] = -1
-		p.W[2] = 200
-		p.W[3] = 0
-		p.W[4] = 0.5
-		p.W[5] = -5
-		p.W[6] = 10
-		p.W[7] = 2.0
-		p.W[8] = 100.0
-		p.W[9] = -1.0
-		p.W[10] = 10.0
-		p.W[11] = -0.5
-		p.W[12] = 1.0
-		p.W[13] = -1.0
-		p.W[14] = -0.5
-		p.W[15] = -1.0
-		p.W[16] = 0.5
-		p.W[20] = 5.0
-		clipParameters(&p)
-		expected := [21][2]float64{
-			{0.001, 100.0},
-			{0.001, 100.0},
-			{0.001, 100.0},
-			{0.001, 100.0},
-			{1.0, 10.0},
-			{0.001, 4.0},
-			{0.001, 4.0},
-			{0.001, 0.75},
-			{0.0, 4.5},
-			{0.0, 0.8},
-			{0.001, 3.5},
-			{0.001, 5.0},
-			{0.001, 0.25},
-			{0.001, 0.9},
-			{0.0, 4.0},
-			{0.0, 1.0},
-			{1.0, 6.0},
-			{0.0, 2.0},
-			{0.0, 2.0},
-			{0.01, 0.8},
-			{0.1, 0.8},
-		}
-		for i := range p.W {
-			if p.W[i] < expected[i][0] || p.W[i] > expected[i][1] {
-				t.Errorf("W[%d]=%v outside range [%v, %v]", i, p.W[i], expected[i][0], expected[i][1])
-			}
-		}
-		if p.W[0] != 0.001 {
-			t.Errorf("W[0]: expected 0.001, got=%v", p.W[0])
-		}
-		if p.W[4] != 1.0 {
-			t.Errorf("W[4]: expected 1.0 (min), got=%v", p.W[4])
-		}
-		if p.W[8] != 4.5 {
-			t.Errorf("W[8]: expected 4.5 (max), got=%v", p.W[8])
-		}
-		if p.W[20] != 0.8 {
-			t.Errorf("W[20]: expected 0.8 (max), got=%v", p.W[20])
-		}
-	})
-
-	t.Run("W20 lower bound raised to 0.1", func(t *testing.T) {
-		p := DefaultParam()
-		p.W[20] = 0
-		clipParameters(&p)
-		if p.W[20] != 0.1 {
-			t.Errorf("W[20]: expected 0.1 (min), got=%v", p.W[20])
-		}
-	})
-
-	t.Run("W19 upper bound clipped to 0.8", func(t *testing.T) {
-		p := DefaultParam()
-		p.W[19] = 1.5
-		clipParameters(&p)
-		if p.W[19] != 0.8 {
-			t.Errorf("W[19]: expected 0.8 (max), got=%v", p.W[19])
-		}
-	})
-
-	t.Run("W17 W18 dynamic ceiling with multiple relearning steps", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{10, 20}
-		originalW17 := p.W[17]
-		originalW18 := p.W[18]
-		clipParameters(&p)
-		if p.W[17] > 2.0 {
-			t.Errorf("W[17]: expected <= 2.0, got=%v", p.W[17])
-		}
-		if p.W[18] > 2.0 {
-			t.Errorf("W[18]: expected <= 2.0, got=%v", p.W[18])
-		}
-		if p.W[17] > originalW17+1e-8 || p.W[18] > originalW18+1e-8 {
-			t.Errorf("W[17]/W[18] should not increase: W[17]=%v (was %v), W[18]=%v (was %v)", p.W[17], originalW17, p.W[18], originalW18)
-		}
-	})
-
-	t.Run("W19 lower bound depends on EnableShortTerm", func(t *testing.T) {
-		p := DefaultParam()
-		p.EnableShortTerm = true
-		p.W[19] = 0.001
-		clipParameters(&p)
-		if p.W[19] < 0.01 {
-			t.Errorf("W[19]: expected >= 0.01 with EnableShortTerm, got=%v", p.W[19])
-		}
-
-		p2 := DefaultParam()
-		p2.EnableShortTerm = false
-		p2.W[19] = 0.001
-		clipParameters(&p2)
-		if p2.W[19] != 0.001 {
-			t.Errorf("W[19]: expected 0.001 without EnableShortTerm, got=%v", p2.W[19])
-		}
-	})
-
-	t.Run("single relearning step uses default ceiling", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{10}
-		p.W[17] = 1.5
-		p.W[18] = 1.5
-		clipParameters(&p)
-		if p.W[17] != 1.5 {
-			t.Errorf("W[17]: expected unchanged 1.5, got=%v", p.W[17])
-		}
-		if p.W[18] != 1.5 {
-			t.Errorf("W[18]: expected unchanged 1.5, got=%v", p.W[18])
-		}
-	})
-
-	t.Run("zero relearning steps uses default ceiling", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{}
-		p.W[17] = 1.5
-		p.W[18] = 1.5
-		clipParameters(&p)
-		if p.W[17] != 1.5 {
-			t.Errorf("W[17]: expected unchanged 1.5, got=%v", p.W[17])
-		}
-	})
-
-	t.Run("dynamic ceiling lower clamp at 0.01", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{10, 20}
-		p.W[11] = 5.0
-		p.W[13] = 0.9
-		p.W[14] = 4.0
-		p.W[17] = 1.5
-		p.W[18] = 1.5
-		clipParameters(&p)
-		if p.W[17] < 0.01-1e-10 || p.W[18] < 0.01-1e-10 {
-			t.Errorf("W[17]/W[18]: expected >= 0.01 (clamped ceiling), got=%v/%v", p.W[17], p.W[18])
-		}
-		if p.W[17] > 2.0+1e-10 || p.W[18] > 2.0+1e-10 {
-			t.Errorf("W[17]/W[18]: expected <= 2.0, got=%v/%v", p.W[17], p.W[18])
-		}
-	})
-}
-
-func TestValidateRequestRetention(t *testing.T) {
-	tests := []struct {
-		name  string
-		value float64
-		ok    bool
-	}{
-		{"valid 0.9", 0.9, true},
-		{"valid 0.5", 0.5, true},
-		{"valid 1.0", 1.0, true},
-		{"valid 0.01", 0.01, true},
-		{"zero", 0, false},
-		{"negative", -0.5, false},
-		{"above 1", 1.5, false},
-		{"NaN", math.NaN(), false},
-		{"Inf", math.Inf(1), false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			p := DefaultParam()
-			p.RequestRetention = tt.value
-			err := p.Validate()
-			if (err == nil) != tt.ok {
-				t.Errorf("RequestRetention=%v: expected ok=%v, got err=%v", tt.value, tt.ok, err)
-			}
-		})
-	}
-}
-
-func TestValidateMaximumInterval(t *testing.T) {
-	tests := []struct {
-		name  string
-		value float64
-		ok    bool
-	}{
-		{"valid 36500", 36500, true},
-		{"valid 1", 1, true},
-		{"zero", 0, false},
-		{"negative", -10, false},
-		{"above max", 36501, false},
-		{"NaN", math.NaN(), false},
-		{"Inf", math.Inf(1), false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			p := DefaultParam()
-			p.MaximumInterval = tt.value
-			err := p.Validate()
-			if (err == nil) != tt.ok {
-				t.Errorf("MaximumInterval=%v: expected ok=%v, got err=%v", tt.value, tt.ok, err)
-			}
-		})
-	}
-}
-
-func TestValidateLearningSteps(t *testing.T) {
-	tests := []struct {
-		name  string
-		value []float64
-		ok    bool
-	}{
-		{"valid default", []float64{1, 10}, true},
-		{"valid empty", []float64{}, true},
-		{"valid zero step", []float64{0}, true},
-		{"negative step", []float64{-1}, false},
-		{"NaN step", []float64{math.NaN()}, false},
-		{"Inf step", []float64{math.Inf(1)}, false},
-		{"mixed valid and invalid", []float64{1, -5}, false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			p := DefaultParam()
-			p.LearningSteps = tt.value
-			err := p.Validate()
-			if (err == nil) != tt.ok {
-				t.Errorf("LearningSteps=%v: expected ok=%v, got err=%v", tt.value, tt.ok, err)
-			}
-		})
-	}
-}
-
-func TestValidateRelearningSteps(t *testing.T) {
-	tests := []struct {
-		name  string
-		value []float64
-		ok    bool
-	}{
-		{"valid default", []float64{10}, true},
-		{"valid empty", []float64{}, true},
-		{"valid zero step", []float64{0}, true},
-		{"negative step", []float64{-5}, false},
-		{"NaN step", []float64{math.NaN()}, false},
-		{"Inf step", []float64{math.Inf(1)}, false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			p := DefaultParam()
-			p.RelearningSteps = tt.value
-			err := p.Validate()
-			if (err == nil) != tt.ok {
-				t.Errorf("RelearningSteps=%v: expected ok=%v, got err=%v", tt.value, tt.ok, err)
-			}
-		})
-	}
-}
-
-func TestNewFSRSClipsWeights(t *testing.T) {
-	p := DefaultParam()
-	p.W[0] = -999
-	p.W[20] = 5.0
-	f := NewFSRS(p)
-	if f.W[0] != 0.001 {
-		t.Errorf("W[0]: expected 0.001 after NewFSRS, got=%v", f.W[0])
-	}
-	if f.W[20] != 0.8 {
-		t.Errorf("W[20]: expected 0.8 after NewFSRS, got=%v", f.W[20])
-	}
-}
-
-func TestNewFSRSDynamicCeiling(t *testing.T) {
-	t.Run("multiple relearning steps clips W17 W18", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{10, 20}
-		p.W[17] = 2.5
-		p.W[18] = 2.5
-		f := NewFSRS(p)
-
-		// w11=1.4835, w13=0.2629, w14=1.6483
-		// value = -(ln(1.4835) + ln(2^0.2629-1) + 1.6483*0.3) / 2 ≈ 0.3601
-		// ceiling = sqrt(0.3601) ≈ 0.6001
-		expectedCeiling := 0.6001
-		tolerance := 1e-2
-		if math.Abs(f.W[17]-expectedCeiling) > tolerance {
-			t.Errorf("W[17]: expected ceiling ≈ %v (with sqrt), got=%v", expectedCeiling, f.W[17])
-		}
-		if math.Abs(f.W[18]-expectedCeiling) > tolerance {
-			t.Errorf("W[18]: expected ceiling ≈ %v (with sqrt), got=%v", expectedCeiling, f.W[18])
-		}
-	})
-
-	t.Run("NaN weights produce finite ceiling", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{10, 20}
-		p.W[11] = -1
-		p.W[17] = 1.5
-		p.W[18] = 1.5
-		f := NewFSRS(p)
-		if math.IsNaN(f.W[17]) || math.IsInf(f.W[17], 0) {
-			t.Errorf("W[17]: expected finite after NewFSRS with negative W[11], got=%v", f.W[17])
-		}
-		if math.IsNaN(f.W[18]) || math.IsInf(f.W[18], 0) {
-			t.Errorf("W[18]: expected finite after NewFSRS with negative W[11], got=%v", f.W[18])
-		}
-		// W[11]=-1 clamped to 0.001 -> value ≈ 4.011 -> sqrt ≈ 2.003 -> clamped to 2.0
-		if f.W[17] > 2.0 {
-			t.Errorf("W[17]: expected <= 2.0, got=%v", f.W[17])
-		}
-	})
-
-	t.Run("single relearning step uses default ceiling of 2.0", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{10}
-		p.W[17] = 2.5
-		p.W[18] = 2.5
-		f := NewFSRS(p)
-
-		if f.W[17] > 2.0 {
-			t.Errorf("W[17]: expected <= 2.0 with single step, got=%v", f.W[17])
-		}
-		if f.W[18] > 2.0 {
-			t.Errorf("W[18]: expected <= 2.0 with single step, got=%v", f.W[18])
-		}
-	})
-
-	t.Run("negative value produces finite ceiling via sqrt guard", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{10, 20}
-		p.W[11] = 5.0
-		p.W[13] = 0.9
-		p.W[14] = 4.0
-		p.W[17] = 1.5
-		p.W[18] = 1.5
-		f := NewFSRS(p)
-
-		// value = -(ln(5) + ln(2^0.9-1) + 4*0.3) / 2 ≈ -1.333
-		// math.Sqrt(math.Max(-1.333, 0)) = sqrt(0) = 0
-		// clamp(0, 0.01, 2.0) = 0.01
-		if math.IsNaN(f.W[17]) || math.IsInf(f.W[17], 0) {
-			t.Errorf("W[17]: expected finite for negative value, got=%v", f.W[17])
-		}
-		if math.IsNaN(f.W[18]) || math.IsInf(f.W[18], 0) {
-			t.Errorf("W[18]: expected finite for negative value, got=%v", f.W[18])
-		}
-		if f.W[17] < 0 || f.W[18] < 0 {
-			t.Errorf("W[17]=%v, W[18]=%v: expected non-negative", f.W[17], f.W[18])
-		}
-	})
-
-	t.Run("W17/W18 ceiling with multiple relearning steps applies square root", func(t *testing.T) {
-		p := DefaultParam()
-		p.RelearningSteps = []float64{10, 20, 30, 40} // len = 4
-		p.W[11] = 0.1
-		p.W[13] = 0.1
-		p.W[14] = 0.5
-		p.W[17] = 2.0
-		p.W[18] = 2.0
-		f := NewFSRS(p)
-
-		// Manual calculation:
-		// w11 = 0.1 -> ln(0.1) ≈ -2.30258509
-		// w13 = 0.1 -> ln(2^0.1 - 1) = ln(1.07177346 - 1) = ln(0.07177346) ≈ -2.63422967
-		// w14 = 0.5 -> w14 * 0.3 = 0.15
-		// sum = -2.30258509 - 2.63422967 + 0.15 = -4.78681476
-		// value = -sum / 4 = 1.19670369
-		// expected ceiling = sqrt(value) ≈ 1.0939395
-		// Without sqrt, the ceiling would be 1.19670369.
-		expectedCeiling := 1.0939395
-		tolerance := 1e-5
-		if math.Abs(f.W[17]-expectedCeiling) > tolerance {
-			t.Errorf("W[17]: expected ceiling ≈ %v (with sqrt), got=%v", expectedCeiling, f.W[17])
-		}
-		if math.Abs(f.W[18]-expectedCeiling) > tolerance {
-			t.Errorf("W[18]: expected ceiling ≈ %v (with sqrt), got=%v", expectedCeiling, f.W[18])
-		}
-
-		// Verify mathematical constraint: w17 * w18 <= value
-		rawValue := 1.19670369
-		constraintTolerance := 1e-4
-		if f.W[17]*f.W[18] > rawValue+constraintTolerance {
-			t.Errorf("W[17]*W[18] = %v, expected <= %v (raw value)", f.W[17]*f.W[18], rawValue)
-		}
-	})
-}
-
-func TestNewFSRSW19EnableShortTerm(t *testing.T) {
-	t.Run("EnableShortTerm true raises W19 to 0.01", func(t *testing.T) {
-		p := DefaultParam()
-		p.EnableShortTerm = true
-		p.W[19] = 0.001
-		f := NewFSRS(p)
-		if f.W[19] < 0.01 {
-			t.Errorf("W[19]: expected >= 0.01 with EnableShortTerm, got=%v", f.W[19])
-		}
-	})
-
-	t.Run("EnableShortTerm false allows W19 below 0.01", func(t *testing.T) {
-		p := DefaultParam()
-		p.EnableShortTerm = false
-		p.W[19] = 0.001
-		f := NewFSRS(p)
-		if f.W[19] != 0.001 {
-			t.Errorf("W[19]: expected 0.001 without EnableShortTerm, got=%v", f.W[19])
-		}
-	})
-}
-
-func TestRetrievability(t *testing.T) {
-	retrievabilityList := []float64{}
-	fsrs := NewFSRS(DefaultParam())
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-	r, err := fsrs.Retrievability(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	retrievabilityList = append(retrievabilityList, roundFloat(r, 4))
-	for range 3 {
-		{
-			rec, err := fsrs.Next(card, now, Good)
-			if err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-			card = rec.Card
-		}
-		now = card.Due
-		r, err := fsrs.Retrievability(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		retrievabilityList = append(retrievabilityList, roundFloat(r, 4))
-	}
-	wantRetrievabilityList := []float64{0, 1, 0.9095, 0.8998}
-	if !reflect.DeepEqual(retrievabilityList, wantRetrievabilityList) {
-		t.Errorf("expected:%v, got:%v", wantRetrievabilityList, retrievabilityList)
-	}
-}
-
-func TestGetRetrievabilityBackwardCompat(t *testing.T) {
-	f := NewFSRS(DefaultParam())
-	card := Card{LastReview: time.Now().Add(-24 * time.Hour), Stability: 1, State: Review}
-	now := card.LastReview.Add(24 * time.Hour)
-
-	got, err := f.GetRetrievability(card, now)
-	if err != nil {
-		t.Fatalf("deprecated GetRetrievability should still work: %v", err)
-	}
-	expected, err := f.Retrievability(card, now)
-	if err != nil {
-		t.Fatalf("Retrievability error: %v", err)
-	}
-	if got != expected {
-		t.Errorf("GetRetrievability = %v, want %v (should match Retrievability)", got, expected)
-	}
-}
-
-func TestDateDiffRawMidnightCrossover(t *testing.T) {
-	last := time.Date(2032, 1, 15, 23, 0, 0, 0, time.UTC)
-	now := time.Date(2032, 1, 16, 1, 0, 0, 0, time.UTC)
-
-	got := dateDiffRaw(last, now)
-	if got != 0 {
-		t.Errorf("dateDiffRaw(midnight crossover) = %v, want 0 (only 2h elapsed)", got)
-	}
-
-	gotCal := dateDiffInDays(last, now)
-	if gotCal != 1 {
-		t.Errorf("dateDiffInDays(midnight crossover) = %d, want 1 (calendar day)", gotCal)
-	}
-}
-
-func TestDateDiffRawFullDay(t *testing.T) {
-	last := time.Date(2032, 1, 15, 10, 0, 0, 0, time.UTC)
-	now := time.Date(2032, 1, 16, 12, 0, 0, 0, time.UTC)
-
-	got := dateDiffRaw(last, now)
-	if got != 1 {
-		t.Errorf("dateDiffRaw(26h) = %v, want 1", got)
-	}
-}
-
-func TestRetrievabilityRawVsCalendar(t *testing.T) {
-	f := NewFSRS(DefaultParam())
-	card := Card{
-		Stability:  5.0,
-		Difficulty: 5.0,
-		State:      Review,
-		LastReview: time.Date(2032, 1, 15, 23, 0, 0, 0, time.UTC),
-	}
-
-	now := time.Date(2032, 1, 16, 1, 0, 0, 0, time.UTC)
-
-	r, err := f.Retrievability(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if r != 1.0 {
-		t.Errorf("Retrievability(midnight crossover, 2h) = %v, want 1.0 (t=0, no decay)", r)
-	}
-}
-
-func TestRetrievabilityReversedDates(t *testing.T) {
-	f := NewFSRS(DefaultParam())
-	card := Card{
-		Stability:  5.0,
-		Difficulty: 5.0,
-		State:      Review,
-		LastReview: time.Date(2032, 1, 16, 12, 0, 0, 0, time.UTC),
-	}
-
-	now := time.Date(2032, 1, 15, 12, 0, 0, 0, time.UTC)
-
-	r, _ := f.Retrievability(card, now)
-	if r != 1.0 {
-		t.Errorf("Retrievability(reversed dates) = %v, want 1.0 (t clamped to 0)", r)
-	}
-}
-
-func TestDecayAndFactorDerivedFromW20(t *testing.T) {
-	p := DefaultParam()
-	p.W[20] = 0.2
-	p.Decay = -0.5
-	p.Factor = math.Pow(0.9, 1.0/p.Decay) - 1.0
-
-	got := p.ForgettingCurve(1, 1)
-	wantDecay := -p.W[20]
-	wantFactor := math.Pow(0.9, 1.0/wantDecay) - 1.0
-	want := math.Pow(1+wantFactor, wantDecay)
-
-	if math.Abs(got-want) > 1e-12 {
-		t.Fatalf("ForgettingCurve should use W[20], got=%v want=%v", got, want)
-	}
-}
-
-func TestInvalidW20ValidationAndFallback(t *testing.T) {
-	p := DefaultParam()
-	p.W[20] = 0
-	if err := p.Validate(); err == nil {
-		t.Fatal("expected validation error for invalid W[20]")
-	}
-
-	gotDecay, gotFactor := p.decayAndFactor()
-	wantDecay, wantFactor := defaultDecayAndFactor()
-	if gotDecay != wantDecay || gotFactor != wantFactor {
-		t.Fatalf("decayAndFactor should fallback to defaults, got=(%v,%v) want=(%v,%v)", gotDecay, gotFactor, wantDecay, wantFactor)
-	}
-
-	pDefault := DefaultParam()
-	if got, want := p.nextInterval(1, 0), pDefault.nextInterval(1, 0); got != want {
-		t.Fatalf("nextInterval should fallback safely, got=%v want=%v", got, want)
-	}
-}
-
-func TestStabilityIsClamped(t *testing.T) {
-	p := DefaultParam()
-	p.W[0] = 1e9
-
-	if got := p.initStability(Again); got != sMax {
-		t.Fatalf("initStability should clamp to sMax, got=%v", got)
-	}
-
-	if got := p.shortTermStability(1e9, Good); got != sMax {
-		t.Fatalf("shortTermStability should clamp to sMax, got=%v", got)
-	}
-}
-
 func BenchmarkRepeat(b *testing.B) {
 	f := NewFSRS(DefaultParam())
 	card := NewCard()
@@ -1005,74 +279,6 @@ func BenchmarkRetrievability(b *testing.B) {
 	}
 	for b.Loop() {
 		_, _ = f.Retrievability(card, now)
-	}
-}
-
-func BenchmarkBasicSchedulerFullSequence(b *testing.B) {
-	f := NewFSRS(DefaultParam())
-	ratings := []Rating{Good, Good, Good, Good, Good, Good, Again, Again, Good, Good, Good, Good, Good}
-	for b.Loop() {
-		card := NewCard()
-		now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-		for _, rating := range ratings {
-			schedCards, err := f.Repeat(card, now)
-			if err != nil {
-				b.Fatalf("unexpected error: %v", err)
-			}
-			record := schedCards[rating]
-			card = record.Card
-			now = card.Due
-		}
-	}
-}
-
-func BenchmarkLongTermSchedulerFullSequence(b *testing.B) {
-	p := DefaultParam()
-	p.EnableShortTerm = false
-	f := NewFSRS(p)
-	ratings := []Rating{Good, Good, Good, Good, Good, Good, Again, Again, Good, Good, Good, Good, Good}
-	for b.Loop() {
-		card := NewCard()
-		now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-		for _, rating := range ratings {
-			schedCards2, err := f.Repeat(card, now)
-			if err != nil {
-				b.Fatalf("unexpected error: %v", err)
-			}
-			record := schedCards2[rating]
-			card = record.Card
-			now = card.Due
-		}
-	}
-}
-
-func BenchmarkNextStates(b *testing.B) {
-	p := DefaultParam()
-	current := &MemoryState{Stability: 51.344814, Difficulty: 7.005062}
-	for b.Loop() {
-		p.NextStates(current, 0.9, 21)
-	}
-}
-
-func BenchmarkNextStates_NewCard(b *testing.B) {
-	p := DefaultParam()
-	for b.Loop() {
-		p.NextStates(nil, 0.9, 0)
-	}
-}
-
-func BenchmarkNextState(b *testing.B) {
-	p := DefaultParam()
-	current := &MemoryState{Stability: 51.344814, Difficulty: 7.005062}
-	for b.Loop() {
-		p.NextState(current, 0.9, 21, Good)
-	}
-}
-
-func BenchmarkForgettingCurve(b *testing.B) {
-	p := DefaultParam()
-	for b.Loop() {
-		p.ForgettingCurve(21, 51.344814)
 	}
 }
 
@@ -1112,3888 +318,5 @@ func TestNewFSRSFallbacksOnInvalidMaximumInterval(t *testing.T) {
 	defaults := DefaultParam()
 	if fsrs.MaximumInterval != defaults.MaximumInterval {
 		t.Errorf("expected default MaximumInterval after fallback, got=%v", fsrs.MaximumInterval)
-	}
-}
-
-func TestNewStateStepBased(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	againCard := schedulingCards[Again].Card
-	if againCard.State != Learning {
-		t.Errorf("Again on new card should be Learning, got=%v", againCard.State)
-	}
-	if againCard.ScheduledDays != 0 {
-		t.Errorf("Again on new card should have ScheduledDays=0, got=%v", againCard.ScheduledDays)
-	}
-	if againCard.RemainingSteps != len(p.LearningSteps) {
-		t.Errorf("Again should have RemainingSteps=%d, got=%d", len(p.LearningSteps), againCard.RemainingSteps)
-	}
-
-	hardCard := schedulingCards[Hard].Card
-	if hardCard.State != Learning {
-		t.Errorf("Hard on new card should be Learning (step-based), got=%v", hardCard.State)
-	}
-	if hardCard.ScheduledDays != 0 {
-		t.Errorf("Hard on new card should have ScheduledDays=0, got=%v", hardCard.ScheduledDays)
-	}
-
-	goodCard := schedulingCards[Good].Card
-	if goodCard.State != Learning {
-		t.Errorf("Good on new card should go to Learning (step 1 of 2), got=%v", goodCard.State)
-	}
-	if goodCard.RemainingSteps != 1 {
-		t.Errorf("Good should have RemainingSteps=1, got=%d", goodCard.RemainingSteps)
-	}
-
-	easyCard := schedulingCards[Easy].Card
-	if easyCard.State != Review {
-		t.Errorf("Easy on new card should graduate to Review, got=%v", easyCard.State)
-	}
-}
-
-func TestLearningStateThresholdBranching(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	againCard := schedulingCards[Again].Card
-	now = againCard.Due
-	schedulingCards, err = fsrs.Repeat(againCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	againAgainCard := schedulingCards[Again].Card
-	if againAgainCard.State != Learning {
-		t.Errorf("Again on learning card should stay in Learning (short interval), got=%v", againAgainCard.State)
-	}
-
-	hardCard := schedulingCards[Hard].Card
-	if hardCard.State != Learning {
-		t.Errorf("Hard on learning card should stay in Learning, got=%v", hardCard.State)
-	}
-
-	goodCard := schedulingCards[Good].Card
-	if goodCard.State != Learning {
-		t.Errorf("Good on learning card should stay in Learning (has remaining step), got=%v", goodCard.State)
-	}
-	if goodCard.ScheduledDays != 0 {
-		t.Errorf("Good on learning card should have ScheduledDays=0, got=%v", goodCard.ScheduledDays)
-	}
-
-	easyCard := schedulingCards[Easy].Card
-	if easyCard.State != Review {
-		t.Errorf("Easy on learning card should graduate to Review, got=%v", easyCard.State)
-	}
-}
-
-func TestLearningStateEasyConstraintWhenGoodGraduates(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-	now = againCard.Due
-	schedulingCards, err = fsrs.Repeat(againCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	goodInterval := schedulingCards[Good].Card.ScheduledDays
-	easyInterval := schedulingCards[Easy].Card.ScheduledDays
-
-	if easyInterval <= goodInterval {
-		t.Errorf("Easy interval (%v) should be > Good interval (%v)",
-			easyInterval, goodInterval)
-	}
-}
-
-func TestReviewStateAgainAlwaysRelearning(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	now = card.Due
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	now = card.Due
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-
-	if againCard.State != Relearning {
-		t.Errorf("Again on review card should always go to Relearning, got=%v", againCard.State)
-	}
-	if againCard.ScheduledDays != 0 {
-		t.Errorf("Again on review card should have ScheduledDays=0 (short-term), got=%d", againCard.ScheduledDays)
-	}
-	if againCard.RemainingSteps != len(p.RelearningSteps) {
-		t.Errorf("Again should have RemainingSteps=%d, got=%d", len(p.RelearningSteps), againCard.RemainingSteps)
-	}
-
-	expectedDue := now.Add(minutesToDuration(p.RelearningSteps[0]))
-	dueDiff := againCard.Due.Sub(expectedDue)
-	if dueDiff < 0 {
-		dueDiff = -dueDiff
-	}
-	if dueDiff > time.Second {
-		t.Errorf("Again Due should match relearning step, expected=%v got=%v diff=%v",
-			expectedDue, againCard.Due, dueDiff)
-	}
-}
-
-func TestCustomLearningSteps(t *testing.T) {
-	p := DefaultParam()
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	p.LearningSteps = []float64{1, 10, 60}
-	s := p.NewBasicScheduler(card, now)
-	schedulingCards := s.Preview()
-
-	againCard := schedulingCards[Again].Card
-	if againCard.State != Learning {
-		t.Errorf("Again should be Learning with steps, got=%v", againCard.State)
-	}
-	if againCard.RemainingSteps != 3 {
-		t.Errorf("Again should have RemainingSteps=3, got=%d", againCard.RemainingSteps)
-	}
-
-	goodCard := schedulingCards[Good].Card
-	if goodCard.State != Learning {
-		t.Errorf("Good should stay Learning with 3 steps (goes to step 2), got=%v", goodCard.State)
-	}
-
-	easyCard := schedulingCards[Easy].Card
-	if easyCard.State != Review {
-		t.Errorf("Easy should always graduate, got=%v", easyCard.State)
-	}
-
-	p.LearningSteps = []float64{}
-	s = p.NewBasicScheduler(card, now)
-	schedulingCards = s.Preview()
-	goodCard = schedulingCards[Good].Card
-	if goodCard.State != Review {
-		t.Errorf("Good should graduate with no steps, got=%v", goodCard.State)
-	}
-}
-
-func TestLearningStateRecallStabilityWhenElapsedDays(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-
-	schedulingCards0, err := fsrs.Repeat(againCard, againCard.Due)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCardImmediate := schedulingCards0[Good].Card
-	stabilityImmediate := goodCardImmediate.Stability
-
-	twoDaysLater := now.Add(2 * 24 * time.Hour)
-	schedulingCards2, err := fsrs.Repeat(againCard, twoDaysLater)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCardDelayed := schedulingCards2[Good].Card
-	stabilityDelayed := goodCardDelayed.Stability
-
-	if stabilityDelayed <= stabilityImmediate {
-		t.Errorf("stability with ElapsedDays>0 (%.4f) should be > stability with ElapsedDays=0 (%.4f)",
-			stabilityDelayed, stabilityImmediate)
-	}
-}
-
-func TestLearningStateHardAndEasyWithElapsedDays(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-
-	twoDaysLater := now.Add(2 * 24 * time.Hour)
-	schedulingCards2, err := fsrs.Repeat(againCard, twoDaysLater)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	hardCardDelayed := schedulingCards2[Hard].Card
-	if hardCardDelayed.Stability <= againCard.Stability {
-		t.Errorf("Hard stability with ElapsedDays=2 (%.4f) should be > initial stability (%.4f)",
-			hardCardDelayed.Stability, againCard.Stability)
-	}
-
-	easyCardDelayed := schedulingCards2[Easy].Card
-	if easyCardDelayed.State != Review {
-		t.Errorf("Easy with ElapsedDays=2 should graduate to Review, got=%v", easyCardDelayed.State)
-	}
-
-	goodCardDelayed := schedulingCards2[Good].Card
-	if easyCardDelayed.ScheduledDays <= goodCardDelayed.ScheduledDays {
-		t.Errorf("Easy interval (%v) should be > Good interval (%v) with ElapsedDays>0",
-			easyCardDelayed.ScheduledDays, goodCardDelayed.ScheduledDays)
-	}
-}
-
-func TestThreeStepProgression(t *testing.T) {
-	p := DefaultParam()
-	p.LearningSteps = []float64{1, 10, 60}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	goodCard := schedulingCards[Good].Card
-	if goodCard.State != Learning {
-		t.Errorf("Good should be Learning at step 1, got=%v", goodCard.State)
-	}
-	if goodCard.RemainingSteps != 2 {
-		t.Errorf("Good should have RS=2, got=%d", goodCard.RemainingSteps)
-	}
-	expectedDue := now.Add(minutesToDuration(10))
-	if goodCard.Due.Sub(expectedDue).Abs() > time.Second {
-		t.Errorf("Good delay should be 10min (step 1), Due=%v expected=%v", goodCard.Due, expectedDue)
-	}
-
-	now = goodCard.Due
-	schedulingCards, err = fsrs.Repeat(goodCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCard2 := schedulingCards[Good].Card
-	if goodCard2.State != Learning {
-		t.Errorf("Good at RS=2 should stay Learning, got=%v", goodCard2.State)
-	}
-	if goodCard2.RemainingSteps != 1 {
-		t.Errorf("Good at RS=2 should have RS=1, got=%d", goodCard2.RemainingSteps)
-	}
-	expectedDue2 := now.Add(minutesToDuration(60))
-	if goodCard2.Due.Sub(expectedDue2).Abs() > time.Second {
-		t.Errorf("Good delay at RS=2 should be 60min (step 2), Due=%v expected=%v", goodCard2.Due, expectedDue2)
-	}
-
-	now = goodCard2.Due
-	schedulingCards, err = fsrs.Repeat(goodCard2, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCard3 := schedulingCards[Good].Card
-	if goodCard3.State != Review {
-		t.Errorf("Good at RS=1 should graduate, got=%v", goodCard3.State)
-	}
-}
-
-func TestEmptyLearningStepsAllGraduate(t *testing.T) {
-	p := DefaultParam()
-	p.LearningSteps = []float64{}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	for _, rating := range []Rating{Again, Hard, Good, Easy} {
-		c := schedulingCards[rating].Card
-		if c.State != Review {
-			t.Errorf("Rating %v with empty steps should graduate to Review, got=%v", rating, c.State)
-		}
-		if c.ScheduledDays == 0 {
-			t.Errorf("Rating %v with empty steps should have interval > 0", rating)
-		}
-	}
-}
-
-func TestHardInterpolation(t *testing.T) {
-	p := DefaultParam()
-
-	p.LearningSteps = []float64{1}
-	hard := hardDelayMinutes(p.LearningSteps)
-	if hard != 2 {
-		t.Errorf("Hard with 1 step [1]: expected 2, got=%v", hard)
-	}
-
-	p.LearningSteps = []float64{1, 10}
-	hard = hardDelayMinutes(p.LearningSteps)
-	if hard != 6 {
-		t.Errorf("Hard with 2 steps [1,10]: expected 6, got=%v", hard)
-	}
-
-	p.LearningSteps = []float64{1, 10, 60}
-	hard = hardDelayMinutes(p.LearningSteps)
-	if hard != 6 {
-		t.Errorf("Hard with 3 steps [1,10,60]: expected 6, got=%v", hard)
-	}
-}
-
-func TestGoodAfterAgainAdvancesStep(t *testing.T) {
-	p := DefaultParam()
-	p.LearningSteps = []float64{1, 10}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-
-	now = againCard.Due
-	schedulingCards, err = fsrs.Repeat(againCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCard := schedulingCards[Good].Card
-
-	expectedDue := now.Add(minutesToDuration(10))
-	if goodCard.Due.Sub(expectedDue).Abs() > time.Second {
-		t.Errorf("Good after Again should have delay=10min (next step), Due=%v expected=%v", goodCard.Due, expectedDue)
-	}
-}
-
-func TestNextStateSingle(t *testing.T) {
-	p := DefaultParam()
-
-	item := p.NextState(nil, 0.9, 0, Good)
-	if item.Memory.Stability != p.initStability(Good) {
-		t.Errorf("NextState new card stability mismatch: got=%v want=%v", item.Memory.Stability, p.initStability(Good))
-	}
-
-	current := &MemoryState{Stability: 5.0, Difficulty: 5.0}
-	item = p.NextState(current, 0.9, 1, Good)
-	if item.Interval < 1 {
-		t.Errorf("NextState interval should be >= 1, got=%v", item.Interval)
-	}
-
-	states := p.NextStates(current, 0.9, 1)
-	if states.Good != item {
-		t.Errorf("NextState(Good) should equal NextStates.Good")
-	}
-}
-
-func TestReviewLogFields(t *testing.T) {
-	t.Run("zero Due card preserves zero in review log", func(t *testing.T) {
-		fsrs := NewFSRS(DefaultParam())
-		card := NewCard(time.Time{})
-		now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-		record, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		log := record.ReviewLog
-
-		if !log.Due.IsZero() {
-			t.Errorf("ReviewLog.Due should be zero for new card with zero Due (pre-review), got=%v", log.Due)
-		}
-		if log.Stability != 0 {
-			t.Errorf("ReviewLog.Stability should be 0 for new card (pre-review), got=%v", log.Stability)
-		}
-		if log.Difficulty != 0 {
-			t.Errorf("ReviewLog.Difficulty should be 0 for new card (pre-review), got=%v", log.Difficulty)
-		}
-		if log.RemainingSteps != 0 {
-			t.Errorf("ReviewLog.RemainingSteps should be 0 for new card (pre-review), got=%d", log.RemainingSteps)
-		}
-	})
-
-	t.Run("default Due card propagates to review log", func(t *testing.T) {
-		fsrs := NewFSRS(DefaultParam())
-		card := NewCard()
-		cardDue := card.Due
-		now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-		record, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		log := record.ReviewLog
-
-		if log.Due.IsZero() {
-			t.Errorf("ReviewLog.Due should be non-zero for card with default Due, got=%v", log.Due)
-		}
-		if log.Due != cardDue {
-			t.Errorf("ReviewLog.Due should equal card.Due (%v), got=%v", cardDue, log.Due)
-		}
-	})
-
-	t.Run("second review log carries previous LastReview", func(t *testing.T) {
-		fsrs := NewFSRS(DefaultParam())
-		card := NewCard(time.Time{})
-		now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-		record, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = record.Card
-		previousLastReview := card.LastReview
-		now = card.Due
-		record, err = fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		log := record.ReviewLog
-
-		if log.Due != previousLastReview {
-			t.Errorf("ReviewLog.Due should be the card's last review time, got=%v want=%v", log.Due, previousLastReview)
-		}
-		if log.Stability == 0 {
-			t.Errorf("ReviewLog.Stability should be non-zero after first review")
-		}
-		if log.Difficulty == 0 {
-			t.Errorf("ReviewLog.Difficulty should be non-zero after first review")
-		}
-		if log.RemainingSteps != 1 {
-			t.Errorf("ReviewLog.RemainingSteps should be 1 (Learning card with 1 step remaining), got=%d", log.RemainingSteps)
-		}
-	})
-
-	t.Run("explicit Due card propagates to review log", func(t *testing.T) {
-		fsrs := NewFSRS(DefaultParam())
-		specificTime := time.Date(2024, 3, 10, 8, 0, 0, 0, time.UTC)
-		card := NewCard(specificTime)
-		now := time.Date(2024, 3, 10, 12, 0, 0, 0, time.UTC)
-
-		record, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		log := record.ReviewLog
-
-		if log.Due != specificTime {
-			t.Errorf("ReviewLog.Due should equal explicit card Due (%v), got=%v", specificTime, log.Due)
-		}
-	})
-}
-
-func TestReviewStateAgainWithEmptyRelearningSteps(t *testing.T) {
-	p := DefaultParam()
-	p.RelearningSteps = []float64{}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	now = card.Due
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	now = card.Due
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-
-	if againCard.State != Review {
-		t.Errorf("Again with empty relearning steps should go to Review, got=%v", againCard.State)
-	}
-	if againCard.ScheduledDays == 0 {
-		t.Errorf("Again with empty relearning steps should have interval > 0")
-	}
-	if againCard.Lapses != 1 {
-		t.Errorf("Again should increment Lapses, got=%d", againCard.Lapses)
-	}
-}
-
-func TestLearningStateRemainingZero(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	card.State = Learning
-	card.Stability = 2.0
-	card.Difficulty = 5.0
-	card.RemainingSteps = 0
-	card.LastReview = time.Date(2022, 11, 29, 12, 0, 0, 0, time.UTC)
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	for _, rating := range []Rating{Again, Hard, Good} {
-		c := schedulingCards[rating].Card
-		if c.State != Review {
-			t.Errorf("Rating %v with RS=0 should graduate to Review, got=%v", rating, c.State)
-		}
-	}
-}
-
-func TestRelearningStepProgression(t *testing.T) {
-	p := DefaultParam()
-	p.RelearningSteps = []float64{10, 60}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	now = card.Due
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	now = card.Due
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-	if againCard.State != Relearning {
-		t.Errorf("Again on review card should go to Relearning, got=%v", againCard.State)
-	}
-	if againCard.RemainingSteps != 2 {
-		t.Errorf("Again should have RS=2, got=%d", againCard.RemainingSteps)
-	}
-
-	now = againCard.Due
-	schedulingCards, err = fsrs.Repeat(againCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCard := schedulingCards[Good].Card
-	if goodCard.State != Relearning {
-		t.Errorf("Good at RS=2 should stay Relearning (step 1 of 2), got=%v", goodCard.State)
-	}
-	if goodCard.RemainingSteps != 1 {
-		t.Errorf("Good at RS=2 should have RS=1, got=%d", goodCard.RemainingSteps)
-	}
-	expectedDue := now.Add(minutesToDuration(60))
-	if goodCard.Due.Sub(expectedDue).Abs() > time.Second {
-		t.Errorf("Good at RS=2 should have delay=60min, Due=%v expected=%v", goodCard.Due, expectedDue)
-	}
-
-	now = goodCard.Due
-	schedulingCards, err = fsrs.Repeat(goodCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCard2 := schedulingCards[Good].Card
-	if goodCard2.State != Review {
-		t.Errorf("Good at RS=1 should graduate to Review, got=%v", goodCard2.State)
-	}
-
-	hardCard := schedulingCards[Hard].Card
-	if hardCard.State != Relearning {
-		t.Errorf("Hard at RS=1 should stay Relearning, got=%v", hardCard.State)
-	}
-}
-
-func TestReviewStateZeroInterval(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	card.State = Review
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-	if againCard.Stability == 0 {
-		t.Errorf("Again with ElapsedDays=0 should use shortTermStability, got stability=0")
-	}
-
-	goodCard := schedulingCards[Good].Card
-	if goodCard.Stability == 0 {
-		t.Errorf("Good with ElapsedDays=0 should use shortTermStability, got stability=0")
-	}
-	if goodCard.State != Review {
-		t.Errorf("Good should be Review, got=%v", goodCard.State)
-	}
-	if goodCard.ScheduledDays == 0 {
-		t.Errorf("Good should have interval > 0, got=%d", goodCard.ScheduledDays)
-	}
-
-	hardCard := schedulingCards[Hard].Card
-	if hardCard.Stability == 0 {
-		t.Errorf("Hard with ElapsedDays=0 should use shortTermStability, got stability=0")
-	}
-	if hardCard.ScheduledDays > goodCard.ScheduledDays {
-		t.Errorf("Hard interval (%d) should be <= Good interval (%d)", hardCard.ScheduledDays, goodCard.ScheduledDays)
-	}
-
-	easyCard := schedulingCards[Easy].Card
-	if easyCard.ScheduledDays <= goodCard.ScheduledDays {
-		t.Errorf("Easy interval (%d) should be > Good interval (%d)", easyCard.ScheduledDays, goodCard.ScheduledDays)
-	}
-}
-
-func TestGoodDelayMinutesBoundary(t *testing.T) {
-	p := DefaultParam()
-
-	p.LearningSteps = []float64{1, 10}
-	delay, ok := goodDelayMinutes(p.LearningSteps, 2)
-	if !ok || delay != 10 {
-		t.Errorf("goodDelayMinutes([1,10], 2): expected (10, true), got=(%v, %v)", delay, ok)
-	}
-
-	delay, ok = goodDelayMinutes(p.LearningSteps, 1)
-	if ok {
-		t.Errorf("goodDelayMinutes([1,10], 1): expected graduation (0, false), got=(%v, %v)", delay, ok)
-	}
-
-	delay, ok = goodDelayMinutes(p.LearningSteps, 0)
-	if ok {
-		t.Errorf("goodDelayMinutes([1,10], 0): expected (0, false), got=(%v, %v)", delay, ok)
-	}
-
-	delay, ok = goodDelayMinutes([]float64{}, 1)
-	if ok {
-		t.Errorf("goodDelayMinutes([], 1): expected (0, false), got=(%v, %v)", delay, ok)
-	}
-
-	p.LearningSteps = []float64{1}
-	delay, ok = goodDelayMinutes(p.LearningSteps, 1)
-	if ok {
-		t.Errorf("goodDelayMinutes([1], 1): expected graduation (0, false), got=(%v, %v)", delay, ok)
-	}
-
-	p.LearningSteps = []float64{1, 10, 60}
-	delay, ok = goodDelayMinutes(p.LearningSteps, 3)
-	if !ok || delay != 10 {
-		t.Errorf("goodDelayMinutes([1,10,60], 3): expected (10, true), got=(%v, %v)", delay, ok)
-	}
-
-	delay, ok = goodDelayMinutes(p.LearningSteps, 2)
-	if !ok || delay != 60 {
-		t.Errorf("goodDelayMinutes([1,10,60], 2): expected (60, true), got=(%v, %v)", delay, ok)
-	}
-
-	delay, ok = goodDelayMinutes(p.LearningSteps, 1)
-	if ok {
-		t.Errorf("goodDelayMinutes([1,10,60], 1): expected graduation (0, false), got=(%v, %v)", delay, ok)
-	}
-}
-
-func TestNextStateElapsedZero(t *testing.T) {
-	p := DefaultParam()
-
-	item := p.NextState(nil, 0.9, 0, Good)
-	if item.Memory.Stability != p.initStability(Good) {
-		t.Errorf("NextState new card: got stability=%v want=%v", item.Memory.Stability, p.initStability(Good))
-	}
-	if item.Memory.Difficulty != constrainDifficulty(p.initDifficulty(Good)) {
-		t.Errorf("NextState new card: got difficulty=%v want=%v", item.Memory.Difficulty, constrainDifficulty(p.initDifficulty(Good)))
-	}
-
-	current := &MemoryState{Stability: 5.0, Difficulty: 5.0}
-	item = p.NextState(current, 0.9, 0, Good)
-	expectedS := p.shortTermStability(5.0, Good)
-	if math.Abs(item.Memory.Stability-expectedS) > 1e-10 {
-		t.Errorf("NextState elapsed=0: got stability=%v want=%v", item.Memory.Stability, expectedS)
-	}
-
-	itemElapsed := p.NextState(current, 0.9, 1, Good)
-	if itemElapsed.Interval < 1 {
-		t.Errorf("NextState elapsed=1: interval should be >= 1, got=%v", itemElapsed.Interval)
-	}
-
-	states := p.NextStates(current, 0.9, 0)
-	if states.Good != item {
-		t.Errorf("NextState(Good, elapsed=0) should equal NextStates.Good")
-	}
-}
-
-func TestReviewLogAllRatings(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	for _, rating := range []Rating{Again, Hard, Good, Easy} {
-		log := schedulingCards[rating].ReviewLog
-		if log.Rating != rating {
-			t.Errorf("ReviewLog.Rating should be %v, got=%v", rating, log.Rating)
-		}
-		if log.State != New {
-			t.Errorf("ReviewLog.State for new card should be New (0), got=%v", log.State)
-		}
-		if log.Stability != 0 {
-			t.Errorf("ReviewLog.Stability for new card should be 0, got=%v", log.Stability)
-		}
-		if log.Difficulty != 0 {
-			t.Errorf("ReviewLog.Difficulty for new card should be 0, got=%v", log.Difficulty)
-		}
-		if log.RemainingSteps != 0 {
-			t.Errorf("ReviewLog.RemainingSteps for new card should be 0, got=%d", log.RemainingSteps)
-		}
-		if log.Review != now {
-			t.Errorf("ReviewLog.Review should be %v, got=%v", now, log.Review)
-		}
-	}
-
-	learningCard := schedulingCards[Good].Card
-	schedulingCards2, err := fsrs.Repeat(learningCard, learningCard.Due)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	for _, rating := range []Rating{Again, Hard, Good, Easy} {
-		log := schedulingCards2[rating].ReviewLog
-		if log.Rating != rating {
-			t.Errorf("Learning ReviewLog.Rating should be %v, got=%v", rating, log.Rating)
-		}
-		if log.Stability == 0 {
-			t.Errorf("Learning ReviewLog.Stability for rating %v should be non-zero", rating)
-		}
-		if log.Difficulty == 0 {
-			t.Errorf("Learning ReviewLog.Difficulty for rating %v should be non-zero", rating)
-		}
-	}
-}
-
-func TestLearningStateHardRemainingStepsInvariant(t *testing.T) {
-	p := DefaultParam()
-	p.LearningSteps = []float64{1, 10, 60}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-	if againCard.RemainingSteps != 3 {
-		t.Errorf("Again should have RS=3, got=%d", againCard.RemainingSteps)
-	}
-
-	now = againCard.Due
-	schedulingCards, err = fsrs.Repeat(againCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	hardCard := schedulingCards[Hard].Card
-	if hardCard.RemainingSteps != 3 {
-		t.Errorf("Hard should keep RS unchanged (3), got=%d", hardCard.RemainingSteps)
-	}
-}
-
-func TestLearningStateAgainResetsRemainingSteps(t *testing.T) {
-	p := DefaultParam()
-	p.LearningSteps = []float64{1, 10, 60}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCard := schedulingCards[Good].Card
-	if goodCard.RemainingSteps != 2 {
-		t.Errorf("Good should have RS=2, got=%d", goodCard.RemainingSteps)
-	}
-
-	now = goodCard.Due
-	schedulingCards, err = fsrs.Repeat(goodCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-	if againCard.RemainingSteps != 3 {
-		t.Errorf("Again should reset RS to full length (3), got=%d", againCard.RemainingSteps)
-	}
-}
-
-func TestHardDelayMinutesIntegration(t *testing.T) {
-	p := DefaultParam()
-	p.LearningSteps = []float64{1, 10}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	hardCard := schedulingCards[Hard].Card
-	if hardCard.State != Learning {
-		t.Errorf("Hard should be Learning, got=%v", hardCard.State)
-	}
-	expectedHardDelay := hardDelayMinutes(p.LearningSteps)
-	expectedDue := now.Add(minutesToDuration(expectedHardDelay))
-	if hardCard.Due.Sub(expectedDue).Abs() > time.Second {
-		t.Errorf("Hard Due should match hardDelayMinutes: due=%v expected=%v", hardCard.Due, expectedDue)
-	}
-	if hardCard.ScheduledDays != 0 {
-		t.Errorf("Hard on new card should have ScheduledDays=0, got=%d", hardCard.ScheduledDays)
-	}
-
-	againCard := schedulingCards[Again].Card
-	now = againCard.Due
-	schedulingCards, err = fsrs.Repeat(againCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	hardCard2 := schedulingCards[Hard].Card
-	if hardCard2.Due.Sub(now.Add(minutesToDuration(expectedHardDelay))).Abs() > time.Second {
-		t.Errorf("Hard in learningState should also match hardDelayMinutes: due=%v", hardCard2.Due)
-	}
-}
-
-func TestDayScaleSteps(t *testing.T) {
-	p := DefaultParam()
-	p.LearningSteps = []float64{1, 10, 1440}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	goodCard := schedulingCards[Good].Card
-	if goodCard.State != Learning {
-		t.Errorf("Good step 1 should be Learning, got=%v", goodCard.State)
-	}
-	if goodCard.ScheduledDays != 0 {
-		t.Errorf("Good step 1 should have ScheduledDays=0, got=%d", goodCard.ScheduledDays)
-	}
-	if goodCard.RemainingSteps != 2 {
-		t.Errorf("Good should have RS=2, got=%d", goodCard.RemainingSteps)
-	}
-
-	now = goodCard.Due
-	schedulingCards, err = fsrs.Repeat(goodCard, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	goodCard2 := schedulingCards[Good].Card
-	if goodCard2.State != Review {
-		t.Errorf("Good step 2 (1440min) should graduate to Review, got=%v", goodCard2.State)
-	}
-	if goodCard2.ScheduledDays != 1 {
-		t.Errorf("Good step 2 (1440min) should have ScheduledDays=1, got=%d", goodCard2.ScheduledDays)
-	}
-	if goodCard2.RemainingSteps != 1 {
-		t.Errorf("Good step 2 should preserve RemainingSteps=1, got=%d", goodCard2.RemainingSteps)
-	}
-	expectedDue := now.Add(1440 * time.Minute)
-	if goodCard2.Due.Sub(expectedDue).Abs() > time.Second {
-		t.Errorf("Good step 2 Due should be now+1440min, due=%v expected=%v", goodCard2.Due, expectedDue)
-	}
-
-	againCard := schedulingCards[Again].Card
-	if againCard.State != Learning {
-		t.Errorf("Again should be Learning (step[0]=1min < 1440), got=%v", againCard.State)
-	}
-}
-
-func TestDayScaleStepRelearning(t *testing.T) {
-	p := DefaultParam()
-	p.RelearningSteps = []float64{1440}
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	now = card.Due
-	{
-		rec, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = rec.Card
-	}
-	now = card.Due
-
-	schedulingCards, err := fsrs.Repeat(card, now)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	againCard := schedulingCards[Again].Card
-	if againCard.State != Review {
-		t.Errorf("Again with 1440min relearning step should graduate to Review, got=%v", againCard.State)
-	}
-	if againCard.ScheduledDays != 1 {
-		t.Errorf("Again should have ScheduledDays=1 (1440min = 1 day), got=%d", againCard.ScheduledDays)
-	}
-	if againCard.Lapses != 1 {
-		t.Errorf("Again should increment Lapses, got=%d", againCard.Lapses)
-	}
-}
-
-func TestForget(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	card := NewCard()
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	ratings := []Rating{Good, Good, Good}
-	for _, r := range ratings {
-		schedulingCards, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[r].Card
-		now = card.Due
-	}
-
-	if card.State == New {
-		t.Fatal("card should not be New before Forget")
-	}
-
-	t.Run("preserves counters", func(t *testing.T) {
-		result := fsrs.Forget(card, now, false)
-		if result.Card.State != New {
-			t.Errorf("expected State=New, got=%v", result.Card.State)
-		}
-		if result.Card.Due != now {
-			t.Errorf("expected Due=now, got=%v", result.Card.Due)
-		}
-		if result.Card.Reps != card.Reps {
-			t.Errorf("expected Reps=%d, got=%d", card.Reps, result.Card.Reps)
-		}
-		if result.Card.Lapses != card.Lapses {
-			t.Errorf("expected Lapses=%d, got=%d", card.Lapses, result.Card.Lapses)
-		}
-		if result.Card.Stability != 0 {
-			t.Errorf("expected Stability=0, got=%v", result.Card.Stability)
-		}
-		if result.Card.Difficulty != 0 {
-			t.Errorf("expected Difficulty=0, got=%v", result.Card.Difficulty)
-		}
-		if result.Card.ScheduledDays != 0 {
-			t.Errorf("expected ScheduledDays=0, got=%d", result.Card.ScheduledDays)
-		}
-		if result.Card.LastReview != card.LastReview {
-			t.Errorf("expected LastReview=%v, got=%v", card.LastReview, result.Card.LastReview)
-		}
-		if result.Card.RemainingSteps != 0 {
-			t.Errorf("expected RemainingSteps=0, got=%d", result.Card.RemainingSteps)
-		}
-		if result.ReviewLog.Rating != Manual {
-			t.Errorf("expected log Rating=Manual, got=%v", result.ReviewLog.Rating)
-		}
-		if result.ReviewLog.State != card.State {
-			t.Errorf("expected log State=%v, got=%v", card.State, result.ReviewLog.State)
-		}
-		if result.ReviewLog.Due != card.Due {
-			t.Errorf("expected log Due=%v, got=%v", card.Due, result.ReviewLog.Due)
-		}
-		if result.ReviewLog.Review != now {
-			t.Errorf("expected log Review=now, got=%v", result.ReviewLog.Review)
-		}
-		if result.ReviewLog.Stability != card.Stability {
-			t.Errorf("expected log Stability=%v, got=%v", card.Stability, result.ReviewLog.Stability)
-		}
-		if result.ReviewLog.Difficulty != card.Difficulty {
-			t.Errorf("expected log Difficulty=%v, got=%v", card.Difficulty, result.ReviewLog.Difficulty)
-		}
-		if result.ReviewLog.RemainingSteps != card.RemainingSteps {
-			t.Errorf("expected log RemainingSteps=%d, got=%d", card.RemainingSteps, result.ReviewLog.RemainingSteps)
-		}
-	})
-
-	t.Run("resets counters", func(t *testing.T) {
-		result := fsrs.Forget(card, now, true)
-		if result.Card.State != New {
-			t.Errorf("expected State=New, got=%v", result.Card.State)
-		}
-		if result.Card.Due != now {
-			t.Errorf("expected Due=now, got=%v", result.Card.Due)
-		}
-		if result.Card.Reps != 0 {
-			t.Errorf("expected Reps=0, got=%d", result.Card.Reps)
-		}
-		if result.Card.Lapses != 0 {
-			t.Errorf("expected Lapses=0, got=%d", result.Card.Lapses)
-		}
-		if result.Card.Stability != 0 {
-			t.Errorf("expected Stability=0, got=%v", result.Card.Stability)
-		}
-		if result.Card.Difficulty != 0 {
-			t.Errorf("expected Difficulty=0, got=%v", result.Card.Difficulty)
-		}
-		if result.Card.ScheduledDays != 0 {
-			t.Errorf("expected ScheduledDays=0, got=%d", result.Card.ScheduledDays)
-		}
-		if result.Card.LastReview != card.LastReview {
-			t.Errorf("expected LastReview=%v, got=%v", card.LastReview, result.Card.LastReview)
-		}
-		if result.Card.RemainingSteps != 0 {
-			t.Errorf("expected RemainingSteps=0, got=%d", result.Card.RemainingSteps)
-		}
-		if result.ReviewLog.Rating != Manual {
-			t.Errorf("expected log Rating=Manual, got=%v", result.ReviewLog.Rating)
-		}
-		if result.ReviewLog.State != card.State {
-			t.Errorf("expected log State=%v, got=%v", card.State, result.ReviewLog.State)
-		}
-		if result.ReviewLog.Due != card.Due {
-			t.Errorf("expected log Due=%v, got=%v", card.Due, result.ReviewLog.Due)
-		}
-		if result.ReviewLog.Stability != card.Stability {
-			t.Errorf("expected log Stability=%v, got=%v", card.Stability, result.ReviewLog.Stability)
-		}
-		if result.ReviewLog.Difficulty != card.Difficulty {
-			t.Errorf("expected log Difficulty=%v, got=%v", card.Difficulty, result.ReviewLog.Difficulty)
-		}
-		if result.ReviewLog.RemainingSteps != card.RemainingSteps {
-			t.Errorf("expected log RemainingSteps=%d, got=%d", card.RemainingSteps, result.ReviewLog.RemainingSteps)
-		}
-	})
-
-	t.Run("already new", func(t *testing.T) {
-		newCard := NewCard(now)
-		result := fsrs.Forget(newCard, now, false)
-		if result.Card.State != New {
-			t.Errorf("expected State=New, got=%v", result.Card.State)
-		}
-		if result.Card.Due != now {
-			t.Errorf("expected Due=now, got=%v", result.Card.Due)
-		}
-		if result.Card.Reps != 0 {
-			t.Errorf("expected Reps=0 for new card, got=%d", result.Card.Reps)
-		}
-		if result.Card.Lapses != 0 {
-			t.Errorf("expected Lapses=0, got=%d", result.Card.Lapses)
-		}
-		if result.Card.Stability != 0 {
-			t.Errorf("expected Stability=0, got=%v", result.Card.Stability)
-		}
-		if result.Card.Difficulty != 0 {
-			t.Errorf("expected Difficulty=0, got=%v", result.Card.Difficulty)
-		}
-		if !result.Card.LastReview.IsZero() {
-			t.Errorf("expected LastReview zero, got=%v", result.Card.LastReview)
-		}
-		if result.ReviewLog.Rating != Manual {
-			t.Errorf("expected log Rating=Manual, got=%v", result.ReviewLog.Rating)
-		}
-		if result.ReviewLog.State != New {
-			t.Errorf("expected log State=New, got=%v", result.ReviewLog.State)
-		}
-		if result.ReviewLog.Due != now {
-			t.Errorf("expected log Due=now, got=%v", result.ReviewLog.Due)
-		}
-		if result.ReviewLog.Review != now {
-			t.Errorf("expected log Review=now, got=%v", result.ReviewLog.Review)
-		}
-		if result.ReviewLog.Stability != 0 {
-			t.Errorf("expected log Stability=0, got=%v", result.ReviewLog.Stability)
-		}
-		if result.ReviewLog.Difficulty != 0 {
-			t.Errorf("expected log Difficulty=0, got=%v", result.ReviewLog.Difficulty)
-		}
-		if result.ReviewLog.ScheduledDays != 0 {
-			t.Errorf("expected log ScheduledDays=0, got=%d", result.ReviewLog.ScheduledDays)
-		}
-		if result.ReviewLog.RemainingSteps != 0 {
-			t.Errorf("expected log RemainingSteps=0, got=%d", result.ReviewLog.RemainingSteps)
-		}
-	})
-
-	t.Run("review state card", func(t *testing.T) {
-		schedulingCards, err := fsrs.Repeat(NewCard(), time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC))
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		reviewCard := schedulingCards[Good].Card
-		if reviewCard.State != Learning && reviewCard.State != Review {
-			t.Fatalf("expected Learning or Review, got=%v", reviewCard.State)
-		}
-		result := fsrs.Forget(reviewCard, reviewCard.Due, false)
-		if result.Card.State != New {
-			t.Errorf("expected State=New, got=%v", result.Card.State)
-		}
-		if result.Card.Due != reviewCard.Due {
-			t.Errorf("expected Due=%v, got=%v", reviewCard.Due, result.Card.Due)
-		}
-		if result.Card.Reps != reviewCard.Reps {
-			t.Errorf("expected Reps=%d, got=%d", reviewCard.Reps, result.Card.Reps)
-		}
-		if result.Card.Lapses != reviewCard.Lapses {
-			t.Errorf("expected Lapses=%d, got=%d", reviewCard.Lapses, result.Card.Lapses)
-		}
-		if result.Card.Stability != 0 {
-			t.Errorf("expected Stability=0, got=%v", result.Card.Stability)
-		}
-		if result.Card.Difficulty != 0 {
-			t.Errorf("expected Difficulty=0, got=%v", result.Card.Difficulty)
-		}
-		if result.Card.LastReview != reviewCard.LastReview {
-			t.Errorf("expected LastReview=%v, got=%v", reviewCard.LastReview, result.Card.LastReview)
-		}
-		if result.ReviewLog.Rating != Manual {
-			t.Errorf("expected log Rating=Manual, got=%v", result.ReviewLog.Rating)
-		}
-		if result.ReviewLog.State != reviewCard.State {
-			t.Errorf("expected log State=%v, got=%v", reviewCard.State, result.ReviewLog.State)
-		}
-		if result.ReviewLog.Due != reviewCard.Due {
-			t.Errorf("expected log Due=%v, got=%v", reviewCard.Due, result.ReviewLog.Due)
-		}
-		if result.ReviewLog.Review != reviewCard.Due {
-			t.Errorf("expected log Review=%v, got=%v", reviewCard.Due, result.ReviewLog.Review)
-		}
-		if result.ReviewLog.Stability != reviewCard.Stability {
-			t.Errorf("expected log Stability=%v, got=%v", reviewCard.Stability, result.ReviewLog.Stability)
-		}
-		if result.ReviewLog.Difficulty != reviewCard.Difficulty {
-			t.Errorf("expected log Difficulty=%v, got=%v", reviewCard.Difficulty, result.ReviewLog.Difficulty)
-		}
-		if result.ReviewLog.RemainingSteps != reviewCard.RemainingSteps {
-			t.Errorf("expected log RemainingSteps=%d, got=%d", reviewCard.RemainingSteps, result.ReviewLog.RemainingSteps)
-		}
-		if result.ReviewLog.ScheduledDays != 0 {
-			t.Errorf("expected log ScheduledDays=0 (Due==now), got=%d", result.ReviewLog.ScheduledDays)
-		}
-	})
-
-	t.Run("log captures pre-forget state", func(t *testing.T) {
-		result := fsrs.Forget(card, now, false)
-		if result.ReviewLog.State != card.State {
-			t.Errorf("expected log State=%v, got=%v", card.State, result.ReviewLog.State)
-		}
-		if result.ReviewLog.Due != card.Due {
-			t.Errorf("expected log Due=%v, got=%v", card.Due, result.ReviewLog.Due)
-		}
-		if result.ReviewLog.Stability != card.Stability {
-			t.Errorf("expected log Stability=%v, got=%v", card.Stability, result.ReviewLog.Stability)
-		}
-		if result.ReviewLog.Difficulty != card.Difficulty {
-			t.Errorf("expected log Difficulty=%v, got=%v", card.Difficulty, result.ReviewLog.Difficulty)
-		}
-		if result.ReviewLog.RemainingSteps != card.RemainingSteps {
-			t.Errorf("expected log RemainingSteps=%d, got=%d", card.RemainingSteps, result.ReviewLog.RemainingSteps)
-		}
-		expectedScheduledDays := uint64(0)
-		if card.State != New {
-			expectedScheduledDays = dateDiffInDays(now, card.Due)
-		}
-		if result.ReviewLog.ScheduledDays != expectedScheduledDays {
-			t.Errorf("expected log ScheduledDays=%d, got=%d", expectedScheduledDays, result.ReviewLog.ScheduledDays)
-		}
-		if result.Card.LastReview != card.LastReview {
-			t.Errorf("expected card LastReview=%v, got=%v", card.LastReview, result.Card.LastReview)
-		}
-		if result.Card.Stability != 0 {
-			t.Errorf("expected card Stability=0, got=%v", result.Card.Stability)
-		}
-		if result.Card.Difficulty != 0 {
-			t.Errorf("expected card Difficulty=0, got=%v", result.Card.Difficulty)
-		}
-		if result.Card.ScheduledDays != 0 {
-			t.Errorf("expected card ScheduledDays=0, got=%d", result.Card.ScheduledDays)
-		}
-		if result.Card.RemainingSteps != 0 {
-			t.Errorf("expected card RemainingSteps=0, got=%d", result.Card.RemainingSteps)
-		}
-	})
-
-	t.Run("log scheduled days with future due", func(t *testing.T) {
-		forgetNow := now.Add(-48 * time.Hour)
-		result := fsrs.Forget(card, forgetNow, false)
-		expectedDays := dateDiffInDays(forgetNow, card.Due)
-		if result.ReviewLog.ScheduledDays != expectedDays {
-			t.Errorf("expected log ScheduledDays=%d, got=%d", expectedDays, result.ReviewLog.ScheduledDays)
-		}
-	})
-}
-
-func TestRollback(t *testing.T) {
-	p := DefaultParam()
-	fsrs := NewFSRS(p)
-	now := time.Date(2022, 11, 29, 12, 30, 0, 0, time.UTC)
-
-	t.Run("restores card after good review", func(t *testing.T) {
-		card := NewCard()
-		schedulingCards, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		result, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		rolledBack, err := fsrs.Rollback(result.Card, result.ReviewLog)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.State != result.ReviewLog.State {
-			t.Errorf("expected State=%v, got=%v", result.ReviewLog.State, rolledBack.State)
-		}
-		if rolledBack.Stability != result.ReviewLog.Stability {
-			t.Errorf("expected Stability=%v, got=%v", result.ReviewLog.Stability, rolledBack.Stability)
-		}
-		if rolledBack.Difficulty != result.ReviewLog.Difficulty {
-			t.Errorf("expected Difficulty=%v, got=%v", result.ReviewLog.Difficulty, rolledBack.Difficulty)
-		}
-		if rolledBack.ScheduledDays != result.ReviewLog.ScheduledDays {
-			t.Errorf("expected ScheduledDays=%d, got=%d", result.ReviewLog.ScheduledDays, rolledBack.ScheduledDays)
-		}
-		if rolledBack.Due != result.ReviewLog.Due {
-			t.Errorf("expected Due=%v, got=%v", result.ReviewLog.Due, rolledBack.Due)
-		}
-		if !rolledBack.LastReview.IsZero() {
-			t.Errorf("expected LastReview zero (New state), got=%v", rolledBack.LastReview)
-		}
-		if rolledBack.Reps != 0 {
-			t.Errorf("expected Reps=0, got=%d", rolledBack.Reps)
-		}
-		if rolledBack.Lapses != 0 {
-			t.Errorf("expected Lapses=0, got=%d", rolledBack.Lapses)
-		}
-		_ = schedulingCards
-	})
-
-	t.Run("decrements lapses on again", func(t *testing.T) {
-		card := NewCard()
-		schedulingCards, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Good].Card
-		now = card.Due
-		schedulingCards, err = fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Good].Card
-		now = card.Due
-		schedulingCards, err = fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Again].Card
-		if card.Lapses == 0 {
-			t.Fatal("expected Lapses > 0 before rollback")
-		}
-		log := schedulingCards[Again].ReviewLog
-		rolledBack, err := fsrs.Rollback(card, log)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.Lapses != card.Lapses-1 {
-			t.Errorf("expected Lapses=%d, got=%d", card.Lapses-1, rolledBack.Lapses)
-		}
-		if rolledBack.Reps != card.Reps-1 {
-			t.Errorf("expected Reps=%d, got=%d", card.Reps-1, rolledBack.Reps)
-		}
-	})
-
-	t.Run("rejects manual rating", func(t *testing.T) {
-		card := NewCard()
-		log := ReviewLog{Rating: Manual}
-		_, err := fsrs.Rollback(card, log)
-		if err == nil {
-			t.Error("expected error for manual rating")
-		}
-		if !errors.Is(err, ErrManualRating) {
-			t.Errorf("expected ErrManualRating, got=%v", err)
-		}
-	})
-
-	t.Run("no underflow on zero reps", func(t *testing.T) {
-		card := NewCard()
-		result, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		rolledBack, err := fsrs.Rollback(result.Card, result.ReviewLog)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.Reps != 0 {
-			t.Errorf("expected Reps=0 (no underflow), got=%d", rolledBack.Reps)
-		}
-	})
-
-	t.Run("lapses unchanged for good rating", func(t *testing.T) {
-		card := NewCard()
-		result, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		rolledBack, err := fsrs.Rollback(result.Card, result.ReviewLog)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.Lapses != 0 {
-			t.Errorf("expected Lapses=0 (no underflow), got=%d", rolledBack.Lapses)
-		}
-	})
-
-	t.Run("restores remaining steps from log", func(t *testing.T) {
-		card := NewCard()
-		schedulingCards, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Again].Card
-		result, err := fsrs.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		rolledBack, err := fsrs.Rollback(result.Card, result.ReviewLog)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.RemainingSteps != result.ReviewLog.RemainingSteps {
-			t.Errorf("expected RemainingSteps=%d (from log), got=%d", result.ReviewLog.RemainingSteps, rolledBack.RemainingSteps)
-		}
-	})
-
-	t.Run("no lapses decrement when again with zero lapses", func(t *testing.T) {
-		card := NewCard()
-		schedulingCards, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Again].Card
-		log := schedulingCards[Again].ReviewLog
-		rolledBack, err := fsrs.Rollback(card, log)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.Lapses != card.Lapses {
-			t.Errorf("expected Lapses=%d (unchanged), got=%d", card.Lapses, rolledBack.Lapses)
-		}
-	})
-
-	t.Run("no lapses decrement for good with positive lapses", func(t *testing.T) {
-		card := NewCard()
-		schedulingCards, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Good].Card
-		now = card.Due
-		schedulingCards, err = fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Good].Card
-		now = card.Due
-		schedulingCards, err = fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Again].Card
-		if card.Lapses == 0 {
-			t.Fatal("expected Lapses > 0 before rollback")
-		}
-		schedulingCards2, err := fsrs.Repeat(card, now)
-		card = schedulingCards2[Good].Card
-		log := schedulingCards2[Good].ReviewLog
-		rolledBack, err := fsrs.Rollback(card, log)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.Lapses != card.Lapses {
-			t.Errorf("expected Lapses=%d (unchanged for Good), got=%d", card.Lapses, rolledBack.Lapses)
-		}
-	})
-
-	t.Run("reps zero on card with zero reps", func(t *testing.T) {
-		card := Card{State: Review, Reps: 0, Lapses: 0}
-		log := ReviewLog{Rating: Good, State: Review}
-		rolledBack, err := fsrs.Rollback(card, log)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.Reps != 0 {
-			t.Errorf("expected Reps=0, got=%d", rolledBack.Reps)
-		}
-	})
-
-	t.Run("manual rating string", func(t *testing.T) {
-		if Manual.String() != "Manual" {
-			t.Errorf("expected Manual.String()=Manual, got=%s", Manual.String())
-		}
-	})
-
-	t.Run("again on relearning preserves lapses", func(t *testing.T) {
-		card := NewCard()
-		schedulingCards, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Good].Card
-		now = card.Due
-		schedulingCards, err = fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Good].Card
-		now = card.Due
-		schedulingCards, err = fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards[Again].Card
-		if card.State != Relearning {
-			t.Fatalf("expected Relearning state, got=%v", card.State)
-		}
-		if card.Lapses == 0 {
-			t.Fatal("expected Lapses > 0")
-		}
-		schedulingCards2, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		card = schedulingCards2[Again].Card
-		log := schedulingCards2[Again].ReviewLog
-		if log.State != Relearning {
-			t.Fatalf("expected log State=Relearning, got=%v", log.State)
-		}
-		rolledBack, err := fsrs.Rollback(card, log)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.Lapses != card.Lapses {
-			t.Errorf("expected Lapses=%d (Again on Relearning does not decrement), got=%d", card.Lapses, rolledBack.Lapses)
-		}
-	})
-
-	t.Run("rejects out of range rating", func(t *testing.T) {
-		card := Card{State: Review, Reps: 1}
-		log := ReviewLog{Rating: Rating(99), State: Review}
-		_, err := fsrs.Rollback(card, log)
-		if err == nil {
-			t.Error("expected error for out-of-range rating")
-		}
-		if !errors.Is(err, ErrInvalidRating) {
-			t.Errorf("expected ErrInvalidRating, got=%v", err)
-		}
-	})
-
-	t.Run("learning state rollback", func(t *testing.T) {
-		card := NewCard()
-		schedulingCards, err := fsrs.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		learningCard := schedulingCards[Again].Card
-		if learningCard.State != Learning {
-			t.Fatalf("expected Learning state, got=%v", learningCard.State)
-		}
-		schedulingCards2, err := fsrs.Repeat(learningCard, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		goodResult := schedulingCards2[Good]
-		rolledBack, err := fsrs.Rollback(goodResult.Card, goodResult.ReviewLog)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if rolledBack.State != Learning {
-			t.Errorf("expected State=Learning, got=%v", rolledBack.State)
-		}
-		if rolledBack.Stability != goodResult.ReviewLog.Stability {
-			t.Errorf("expected Stability=%v, got=%v", goodResult.ReviewLog.Stability, rolledBack.Stability)
-		}
-		if rolledBack.Difficulty != goodResult.ReviewLog.Difficulty {
-			t.Errorf("expected Difficulty=%v, got=%v", goodResult.ReviewLog.Difficulty, rolledBack.Difficulty)
-		}
-		if rolledBack.RemainingSteps != goodResult.ReviewLog.RemainingSteps {
-			t.Errorf("expected RemainingSteps=%d, got=%d", goodResult.ReviewLog.RemainingSteps, rolledBack.RemainingSteps)
-		}
-		if rolledBack.Due != goodResult.ReviewLog.Review {
-			t.Errorf("expected Due=%v, got=%v", goodResult.ReviewLog.Review, rolledBack.Due)
-		}
-		if rolledBack.LastReview != goodResult.ReviewLog.Due {
-			t.Errorf("expected LastReview=%v, got=%v", goodResult.ReviewLog.Due, rolledBack.LastReview)
-		}
-	})
-}
-
-func TestReschedule(t *testing.T) {
-	p := DefaultParam()
-	f := NewFSRS(p)
-
-	t.Run("basic grade replay matches direct replay", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-		card := NewCard()
-		result, err := f.Reschedule(card, reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 4 {
-			t.Fatalf("expected 4 collections, got %d", len(result.Collections))
-		}
-
-		curCard := Card{Due: card.Due}
-		for i, review := range reviews {
-			direct, err := f.Next(curCard, review.Review, review.Rating)
-			if err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-			if !reflect.DeepEqual(result.Collections[i].Card, direct.Card) {
-				t.Errorf("review %d: reschedule card mismatch\n  got:  %+v\n  want: %+v", i, result.Collections[i].Card, direct.Card)
-			}
-			curCard = result.Collections[i].Card
-		}
-	})
-
-	t.Run("basic grade exact values", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		wantIntervals := []uint64{0, 2, 16, 53}
-		for i, want := range wantIntervals {
-			if got := result.Collections[i].Card.ScheduledDays; got != want {
-				t.Errorf("review %d: scheduled_days=%d, want=%d", i, got, want)
-			}
-		}
-
-		wantStability := []float64{2.3065, 2.3065, 16.1880, 52.7633}
-		for i, want := range wantStability {
-			if got := roundFloat(result.Collections[i].Card.Stability, 4); got != want {
-				t.Errorf("review %d: stability=%.10f, want=%.4f", i, result.Collections[i].Card.Stability, want)
-			}
-		}
-
-		wantDifficulty := []float64{2.1181, 2.1112, 2.1043, 2.0975}
-		for i, want := range wantDifficulty {
-			if got := roundFloat(result.Collections[i].Card.Difficulty, 4); got != want {
-				t.Errorf("review %d: difficulty=%.10f, want=%.4f", i, result.Collections[i].Card.Difficulty, want)
-			}
-		}
-
-		if result.RescheduleItem == nil {
-			t.Error("expected non-nil reschedule_item")
-		}
-	})
-
-	t.Run("basic grade long-term", func(t *testing.T) {
-		ltParam := DefaultParam()
-		ltParam.EnableShortTerm = false
-		ltFsrs := NewFSRS(ltParam)
-
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		result, err := ltFsrs.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		wantIntervals := []uint64{3, 3, 16, 53}
-		for i, want := range wantIntervals {
-			if got := result.Collections[i].Card.ScheduledDays; got != want {
-				t.Errorf("review %d: scheduled_days=%d, want=%d", i, got, want)
-			}
-		}
-
-		wantStability := []float64{2.3065, 2.3065, 16.1880, 52.7633}
-		for i, want := range wantStability {
-			if got := roundFloat(result.Collections[i].Card.Stability, 4); got != want {
-				t.Errorf("review %d: stability=%.10f, want=%.4f", i, result.Collections[i].Card.Stability, want)
-			}
-		}
-
-		if result.RescheduleItem == nil {
-			t.Error("expected non-nil reschedule_item")
-		}
-	})
-
-	t.Run("empty reviews", func(t *testing.T) {
-		result, err := f.Reschedule(NewCard(), nil, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 0 {
-			t.Errorf("expected 0 collections, got %d", len(result.Collections))
-		}
-		if result.RescheduleItem != nil {
-			t.Error("expected nil reschedule_item")
-		}
-	})
-
-	t.Run("manual rating with state New", func(t *testing.T) {
-		reviewTime := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: reviewTime},
-			{Rating: Manual, Review: reviewTime.Add(24 * time.Hour), State: StatePtr(New)},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 2 {
-			t.Fatalf("expected 2 collections, got %d", len(result.Collections))
-		}
-
-		lastCard := result.Collections[1].Card
-		if lastCard.State != New {
-			t.Errorf("expected State=New, got=%v", lastCard.State)
-		}
-		if lastCard.Stability != 0 {
-			t.Errorf("expected Stability=0, got=%v", lastCard.Stability)
-		}
-		if lastCard.Difficulty != 0 {
-			t.Errorf("expected Difficulty=0, got=%v", lastCard.Difficulty)
-		}
-		if lastCard.Due.IsZero() {
-			t.Error("expected non-zero Due")
-		}
-
-		lastLog := result.Collections[1].ReviewLog
-		if lastLog.Rating != Manual {
-			t.Errorf("expected log Rating=Manual, got=%v", lastLog.Rating)
-		}
-		if lastLog.State != New {
-			t.Errorf("expected log State=New, got=%v", lastLog.State)
-		}
-		if !lastLog.Due.Equal(reviewTime.Add(24 * time.Hour)) {
-			t.Errorf("expected log Due=reviewed time, got=%v", lastLog.Due)
-		}
-		if lastLog.ScheduledDays != 0 {
-			t.Errorf("expected log ScheduledDays=0, got=%d", lastLog.ScheduledDays)
-		}
-	})
-
-	t.Run("manual rating with state New and explicit Due", func(t *testing.T) {
-		reviewTime := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		manualDue := time.Date(2024, 9, 20, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: reviewTime},
-			{Rating: Manual, Review: reviewTime.Add(24 * time.Hour), State: StatePtr(New), Due: manualDue},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		lastCard := result.Collections[1].Card
-		if !lastCard.Due.Equal(manualDue) {
-			t.Errorf("expected Due=%v (explicit), got=%v", manualDue, lastCard.Due)
-		}
-		if lastCard.ScheduledDays != 6 {
-			t.Errorf("expected ScheduledDays=6, got=%d", lastCard.ScheduledDays)
-		}
-	})
-
-	t.Run("manual rating with state New and Due before Review", func(t *testing.T) {
-		reviewTime := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		dueBeforeReview := time.Date(2024, 9, 10, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: reviewTime},
-			{Rating: Manual, Review: reviewTime.Add(24 * time.Hour), State: StatePtr(New), Due: dueBeforeReview},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		lastCard := result.Collections[1].Card
-		if !lastCard.Due.Equal(dueBeforeReview) {
-			t.Errorf("expected Due=%v, got=%v", dueBeforeReview, lastCard.Due)
-		}
-		if lastCard.ScheduledDays != 0 {
-			t.Errorf("expected ScheduledDays=0 when Due <= Review, got=%d", lastCard.ScheduledDays)
-		}
-	})
-
-	t.Run("manual rating with state Review", func(t *testing.T) {
-		review1 := time.Date(2024, 8, 12, 1, 0, 0, 0, time.UTC)
-		review2 := time.Date(2024, 8, 13, 1, 0, 0, 0, time.UTC)
-		review3 := time.Date(2024, 8, 14, 1, 0, 0, 0, time.UTC)
-		review4 := time.Date(2024, 8, 15, 1, 0, 0, 0, time.UTC)
-		manualDue := time.Date(2024, 9, 4, 17, 0, 0, 0, time.UTC)
-
-		reviews := []ReviewHistory{
-			{Rating: Easy, Review: review1},
-			{Rating: Good, Review: review2},
-			{Rating: Manual, Review: review3, State: StatePtr(Review), Stability: 21.79806877, Difficulty: 3.2828565, Due: manualDue},
-			{Rating: Good, Review: review4},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 4 {
-			t.Fatalf("expected 4 collections, got %d", len(result.Collections))
-		}
-
-		manualCard := result.Collections[2].Card
-		if manualCard.State != Review {
-			t.Errorf("expected State=Review, got=%v", manualCard.State)
-		}
-		if math.Abs(manualCard.Stability-21.79806877) > 1e-6 {
-			t.Errorf("expected Stability=21.79806877, got=%v", manualCard.Stability)
-		}
-		if math.Abs(manualCard.Difficulty-3.2828565) > 1e-6 {
-			t.Errorf("expected Difficulty=3.2828565, got=%v", manualCard.Difficulty)
-		}
-		if !manualCard.Due.Equal(manualDue) {
-			t.Errorf("expected Due=%v, got=%v", manualDue, manualCard.Due)
-		}
-		if manualCard.Reps != 3 {
-			t.Errorf("expected Reps=3, got=%d", manualCard.Reps)
-		}
-		if manualCard.ScheduledDays != 21 {
-			t.Errorf("expected ScheduledDays=21, got=%d", manualCard.ScheduledDays)
-		}
-		if !manualCard.LastReview.Equal(review3) {
-			t.Errorf("expected LastReview=%v, got=%v", review3, manualCard.LastReview)
-		}
-
-		manualLog := result.Collections[2].ReviewLog
-		if manualLog.Rating != Manual {
-			t.Errorf("expected log Rating=Manual, got=%v", manualLog.Rating)
-		}
-		if manualLog.State != Review {
-			t.Errorf("expected log State=Review (pre-review), got=%v", manualLog.State)
-		}
-		if manualLog.ScheduledDays != 13 {
-			t.Errorf("expected log ScheduledDays=13, got=%d", manualLog.ScheduledDays)
-		}
-	})
-
-	t.Run("manual without state returns error", func(t *testing.T) {
-		reviewTime := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{
-			{Rating: Easy, Review: reviewTime},
-			{Rating: Good, Review: reviewTime.Add(24 * time.Hour)},
-			{Rating: Manual, Review: reviewTime.Add(48 * time.Hour)},
-			{Rating: Good, Review: reviewTime.Add(72 * time.Hour)},
-		}
-
-		_, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err == nil {
-			t.Fatal("expected error for manual rating without state")
-		}
-		if !errors.Is(err, ErrManualStateRequired) {
-			t.Errorf("expected ErrManualStateRequired, got=%v", err)
-		}
-	})
-
-	t.Run("manual non-New without due returns error", func(t *testing.T) {
-		reviewTime := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{
-			{Rating: Easy, Review: reviewTime},
-			{Rating: Good, Review: reviewTime.Add(24 * time.Hour)},
-			{Rating: Manual, Review: reviewTime.Add(48 * time.Hour), State: StatePtr(Review)},
-			{Rating: Good, Review: reviewTime.Add(72 * time.Hour)},
-		}
-
-		_, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err == nil {
-			t.Fatal("expected error for manual rating without due")
-		}
-		if !errors.Is(err, ErrManualDueRequired) {
-			t.Errorf("expected ErrManualDueRequired, got=%v", err)
-		}
-	})
-
-	t.Run("reschedule item generated", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		currentCard := Card{
-			Due:       time.Date(2024, 12, 25, 0, 0, 0, 0, time.UTC),
-			State:     Review,
-			Reps:      5,
-			Stability: 50.0,
-		}
-		now := time.Date(2024, 10, 27, 0, 0, 0, 0, time.UTC)
-
-		result, err := f.Reschedule(currentCard, reviews, RescheduleOptions{
-			UpdateMemoryState: true,
-			Now:               now,
-		})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		if result.RescheduleItem == nil {
-			t.Fatal("expected non-nil reschedule_item")
-		}
-		if result.RescheduleItem.ReviewLog.Rating != Manual {
-			t.Errorf("expected log Rating=Manual, got=%v", result.RescheduleItem.ReviewLog.Rating)
-		}
-		if !result.RescheduleItem.Card.LastReview.Equal(now) {
-			t.Errorf("expected card LastReview=now, got=%v", result.RescheduleItem.Card.LastReview)
-		}
-		if result.RescheduleItem.Card.Reps != currentCard.Reps+1 {
-			t.Errorf("expected card Reps=%d, got=%d", currentCard.Reps+1, result.RescheduleItem.Card.Reps)
-		}
-	})
-
-	t.Run("reschedule item null when due matches", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		lastCard := result.Collections[len(result.Collections)-1].Card
-
-		currentCard := lastCard
-		result2, err := f.Reschedule(currentCard, reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if result2.RescheduleItem != nil {
-			t.Error("expected nil reschedule_item when due matches")
-		}
-	})
-
-	t.Run("skip manual filters Manual ratings", func(t *testing.T) {
-		reviewTime := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: reviewTime},
-			{Rating: Manual, Review: reviewTime.Add(24 * time.Hour), State: StatePtr(New)},
-			{Rating: Good, Review: reviewTime.Add(48 * time.Hour)},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{SkipManual: true})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 2 {
-			t.Errorf("expected 2 collections (Manual filtered), got %d", len(result.Collections))
-		}
-	})
-
-	t.Run("skip manual with all-Manual reviews yields empty collections", func(t *testing.T) {
-		r1 := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		r2 := time.Date(2024, 9, 14, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{
-			{Rating: Manual, Review: r1, State: StatePtr(New)},
-			{Rating: Manual, Review: r2, State: StatePtr(New)},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{SkipManual: true})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 0 {
-			t.Errorf("expected 0 collections, got %d", len(result.Collections))
-		}
-		if result.RescheduleItem != nil {
-			t.Error("expected nil RescheduleItem with empty collections")
-		}
-	})
-
-	t.Run("sort reviews by review time", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		sortedReviews := make([]ReviewHistory, len(reviewTimes))
-		sortedTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		for i, rt := range sortedTimes {
-			sortedReviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		resultSorted, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{SortReviews: true})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		resultPreSorted, err := f.Reschedule(NewCard(), sortedReviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		for i := range resultSorted.Collections {
-			if !reflect.DeepEqual(resultSorted.Collections[i].Card, resultPreSorted.Collections[i].Card) {
-				t.Errorf("review %d: sorted result doesn't match pre-sorted", i)
-			}
-		}
-	})
-
-	t.Run("first card override", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		firstDue := time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC)
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{FirstDue: firstDue})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 4 {
-			t.Fatalf("expected 4 collections, got %d", len(result.Collections))
-		}
-	})
-
-	t.Run("forget scenario", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		firstDue := time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC)
-
-		var currentCard Card
-		var historyCards []Card
-		for _, review := range reviews {
-				item, err := f.Next(currentCard, review.Review, review.Rating)
-			if err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-			currentCard = item.Card
-			historyCards = append(historyCards, currentCard)
-		}
-
-		forgetItem := f.Forget(currentCard, time.Date(2024, 10, 27, 0, 0, 0, 0, time.UTC), false)
-		currentCard = forgetItem.Card
-
-		result, err := f.Reschedule(currentCard, reviews, RescheduleOptions{
-			FirstDue:          firstDue,
-			UpdateMemoryState: true,
-			Now:               time.Date(2024, 10, 27, 0, 0, 0, 0, time.UTC),
-		})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		if result.RescheduleItem == nil {
-			t.Fatal("expected non-nil reschedule_item")
-		}
-		lastHistory := historyCards[len(historyCards)-1]
-		if !result.RescheduleItem.Card.Due.Equal(lastHistory.Due) {
-			t.Errorf("expected card Due=%v, got=%v", lastHistory.Due, result.RescheduleItem.Card.Due)
-		}
-		if math.Abs(result.RescheduleItem.Card.Stability-lastHistory.Stability) > 1e-10 {
-			t.Errorf("expected card Stability=%v, got=%v", lastHistory.Stability, result.RescheduleItem.Card.Stability)
-		}
-		if math.Abs(result.RescheduleItem.Card.Difficulty-lastHistory.Difficulty) > 1e-10 {
-			t.Errorf("expected card Difficulty=%v, got=%v", lastHistory.Difficulty, result.RescheduleItem.Card.Difficulty)
-		}
-	})
-
-	t.Run("manual rating without stability uses card values", func(t *testing.T) {
-		review1 := time.Date(2024, 8, 12, 1, 0, 0, 0, time.UTC)
-		review2 := time.Date(2024, 8, 13, 1, 0, 0, 0, time.UTC)
-		review3 := time.Date(2024, 8, 14, 1, 0, 0, 0, time.UTC)
-		manualDue := time.Date(2024, 9, 4, 17, 0, 0, 0, time.UTC)
-
-		reviews := []ReviewHistory{
-			{Rating: Easy, Review: review1},
-			{Rating: Good, Review: review2},
-			{Rating: Manual, Review: review3, State: StatePtr(Review), Due: manualDue},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		manualCard := result.Collections[2].Card
-		prevCard := result.Collections[1].Card
-		if manualCard.Stability != prevCard.Stability {
-			t.Errorf("expected Stability=%v (from previous card), got=%v", prevCard.Stability, manualCard.Stability)
-		}
-		if manualCard.Difficulty != prevCard.Difficulty {
-			t.Errorf("expected Difficulty=%v (from previous card), got=%v", prevCard.Difficulty, manualCard.Difficulty)
-		}
-	})
-
-	t.Run("sort reviews does not mutate caller slice", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-		originalOrder := make([]time.Time, len(reviews))
-		for i, r := range reviews {
-			originalOrder[i] = r.Review
-		}
-
-		_, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{SortReviews: true})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		for i, r := range reviews {
-			if !r.Review.Equal(originalOrder[i]) {
-				t.Errorf("caller slice mutated at index %d: got=%v, want=%v", i, r.Review, originalOrder[i])
-			}
-		}
-	})
-
-	t.Run("sort reviews with skip manual combined", func(t *testing.T) {
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC)},
-			{Rating: Manual, Review: time.Date(2024, 9, 14, 0, 0, 0, 0, time.UTC), State: StatePtr(New)},
-			{Rating: Good, Review: time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{SortReviews: true, SkipManual: true})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 2 {
-			t.Fatalf("expected 2 collections, got %d", len(result.Collections))
-		}
-		firstReview := result.Collections[0].ReviewLog.Review
-		secondReview := result.Collections[1].ReviewLog.Review
-		if !firstReview.Before(secondReview) {
-			t.Errorf("expected sorted order: first=%v should be before second=%v", firstReview, secondReview)
-		}
-	})
-
-	t.Run("single review", func(t *testing.T) {
-		reviewTime := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{{Rating: Good, Review: reviewTime}}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 1 {
-			t.Fatalf("expected 1 collection, got %d", len(result.Collections))
-		}
-		if result.Collections[0].Card.State == New {
-			t.Errorf("expected non-New state after Good review, got=%v", result.Collections[0].Card.State)
-		}
-	})
-
-	t.Run("all manual reviews", func(t *testing.T) {
-		r1 := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		r2 := time.Date(2024, 9, 14, 0, 0, 0, 0, time.UTC)
-		reviews := []ReviewHistory{
-			{Rating: Manual, Review: r1, State: StatePtr(New)},
-			{Rating: Manual, Review: r2, State: StatePtr(New)},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(result.Collections) != 2 {
-			t.Fatalf("expected 2 collections, got %d", len(result.Collections))
-		}
-		for i, c := range result.Collections {
-			if c.Card.State != New {
-				t.Errorf("collection %d: expected State=New, got=%v", i, c.Card.State)
-			}
-		}
-	})
-
-	t.Run("manual rating with Learning state", func(t *testing.T) {
-		r1 := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		r2 := time.Date(2024, 9, 14, 0, 0, 0, 0, time.UTC)
-		manualDue := time.Date(2024, 9, 14, 10, 0, 0, 0, time.UTC)
-
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: r1},
-			{Rating: Manual, Review: r2, State: StatePtr(Learning), Stability: 2.5, Difficulty: 4.0, Due: manualDue},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		manualCard := result.Collections[1].Card
-		if manualCard.State != Learning {
-			t.Errorf("expected State=Learning, got=%v", manualCard.State)
-		}
-		if math.Abs(manualCard.Stability-2.5) > 1e-6 {
-			t.Errorf("expected Stability=2.5, got=%v", manualCard.Stability)
-		}
-		if !manualCard.Due.Equal(manualDue) {
-			t.Errorf("expected Due=%v, got=%v", manualDue, manualCard.Due)
-		}
-	})
-
-	t.Run("manual rating with Relearning state", func(t *testing.T) {
-		r1 := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		r2 := time.Date(2024, 9, 14, 0, 0, 0, 0, time.UTC)
-		r3 := time.Date(2024, 9, 15, 0, 0, 0, 0, time.UTC)
-		manualDue := time.Date(2024, 9, 15, 10, 0, 0, 0, time.UTC)
-
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: r1},
-			{Rating: Manual, Review: r2, State: StatePtr(Review), Stability: 10.0, Difficulty: 5.0, Due: time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC)},
-			{Rating: Manual, Review: r3, State: StatePtr(Relearning), Stability: 5.0, Difficulty: 6.0, Due: manualDue},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		relearnCard := result.Collections[2].Card
-		if relearnCard.State != Relearning {
-			t.Errorf("expected State=Relearning, got=%v", relearnCard.State)
-		}
-		if math.Abs(relearnCard.Stability-5.0) > 1e-6 {
-			t.Errorf("expected Stability=5.0, got=%v", relearnCard.Stability)
-		}
-	})
-
-	t.Run("update memory state false preserves stability", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		now := time.Date(2024, 10, 27, 0, 0, 0, 0, time.UTC)
-		currentCard := Card{Due: now, State: Review, Reps: 3, Stability: 10.0, Difficulty: 5.0}
-
-		result, err := f.Reschedule(currentCard, reviews, RescheduleOptions{
-			UpdateMemoryState: false,
-			Now:               now,
-		})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if result.RescheduleItem == nil {
-			t.Fatal("expected non-nil reschedule_item")
-		}
-		if result.RescheduleItem.Card.Stability != currentCard.Stability {
-			t.Errorf("with UpdateMemoryState=false: expected Stability=%v (unchanged), got=%v", currentCard.Stability, result.RescheduleItem.Card.Stability)
-		}
-		if result.RescheduleItem.Card.Difficulty != currentCard.Difficulty {
-			t.Errorf("with UpdateMemoryState=false: expected Difficulty=%v (unchanged), got=%v", currentCard.Difficulty, result.RescheduleItem.Card.Difficulty)
-		}
-	})
-
-	t.Run("reschedule item for manual-last replay with non-New state", func(t *testing.T) {
-		r1 := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		r2 := time.Date(2024, 9, 14, 0, 0, 0, 0, time.UTC)
-		manualDue := time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC)
-
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: r1},
-			{Rating: Manual, Review: r2, State: StatePtr(Review), Stability: 10.0, Difficulty: 5.0, Due: manualDue},
-		}
-
-		now := time.Date(2024, 10, 27, 0, 0, 0, 0, time.UTC)
-		currentCard := Card{Due: now, State: Review, Reps: 3, Stability: 1.0, Difficulty: 1.0}
-
-		result, err := f.Reschedule(currentCard, reviews, RescheduleOptions{
-			UpdateMemoryState: true,
-			Now:               now,
-		})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if result.RescheduleItem == nil {
-			t.Fatal("expected non-nil reschedule_item for manual-last replay")
-		}
-		if result.RescheduleItem.Card.State != Review {
-			t.Errorf("expected State=Review, got=%v", result.RescheduleItem.Card.State)
-		}
-		if math.Abs(result.RescheduleItem.Card.Stability-10.0) > 1e-6 {
-			t.Errorf("with UpdateMemoryState: expected Stability=10.0, got=%v", result.RescheduleItem.Card.Stability)
-		}
-		if math.Abs(result.RescheduleItem.Card.Difficulty-5.0) > 1e-6 {
-			t.Errorf("with UpdateMemoryState: expected Difficulty=5.0, got=%v", result.RescheduleItem.Card.Difficulty)
-		}
-	})
-
-	t.Run("reschedule item for manual-last New state", func(t *testing.T) {
-		r1 := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		r2 := time.Date(2024, 9, 14, 0, 0, 0, 0, time.UTC)
-
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: r1},
-			{Rating: Manual, Review: r2, State: StatePtr(New)},
-		}
-
-		now := time.Date(2024, 10, 27, 0, 0, 0, 0, time.UTC)
-		currentCard := Card{Due: now, State: Review, Reps: 3, Stability: 10.0, Difficulty: 5.0}
-
-		result, err := f.Reschedule(currentCard, reviews, RescheduleOptions{
-			Now: now,
-		})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if result.RescheduleItem == nil {
-			t.Fatal("expected non-nil reschedule_item when current card Due differs from manual-last New card")
-		}
-		if result.RescheduleItem.Card.State != New {
-			t.Errorf("expected State=New, got=%v", result.RescheduleItem.Card.State)
-		}
-	})
-
-	t.Run("manual due before review produces zero scheduled days", func(t *testing.T) {
-		r1 := time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC)
-		dueBeforeReview := time.Date(2024, 9, 10, 0, 0, 0, 0, time.UTC)
-
-		reviews := []ReviewHistory{
-			{Rating: Good, Review: r1},
-			{Rating: Manual, Review: time.Date(2024, 9, 14, 0, 0, 0, 0, time.UTC), State: StatePtr(Review), Stability: 5.0, Difficulty: 4.0, Due: dueBeforeReview},
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		manualCard := result.Collections[1].Card
-		if manualCard.ScheduledDays != 0 {
-			t.Errorf("expected ScheduledDays=0 when due < review, got=%d", manualCard.ScheduledDays)
-		}
-	})
-
-	t.Run("reschedule item with update memory state", func(t *testing.T) {
-		reviewTimes := []time.Time{
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 13, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 17, 0, 0, 0, 0, time.UTC),
-			time.Date(2024, 9, 28, 0, 0, 0, 0, time.UTC),
-		}
-		reviews := make([]ReviewHistory, len(reviewTimes))
-		for i, rt := range reviewTimes {
-			reviews[i] = ReviewHistory{Rating: Good, Review: rt}
-		}
-
-		result, err := f.Reschedule(NewCard(), reviews, RescheduleOptions{})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		lastCard := result.Collections[len(result.Collections)-1].Card
-
-		now := time.Date(2024, 10, 27, 0, 0, 0, 0, time.UTC)
-		currentCard := Card{Due: now, State: Review, Reps: 3, Stability: 10.0, Difficulty: 5.0}
-
-		result2, err := f.Reschedule(currentCard, reviews, RescheduleOptions{
-			UpdateMemoryState: true,
-			Now:               now,
-		})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if result2.RescheduleItem == nil {
-			t.Fatal("expected non-nil reschedule_item")
-		}
-
-		if math.Abs(result2.RescheduleItem.Card.Stability-lastCard.Stability) > 1e-10 {
-			t.Errorf("with UpdateMemoryState: expected Stability=%v, got=%v", lastCard.Stability, result2.RescheduleItem.Card.Stability)
-		}
-		if math.Abs(result2.RescheduleItem.Card.Difficulty-lastCard.Difficulty) > 1e-10 {
-			t.Errorf("with UpdateMemoryState: expected Difficulty=%v, got=%v", lastCard.Difficulty, result2.RescheduleItem.Card.Difficulty)
-		}
-	})
-}
-
-func TestApplyFuzz(t *testing.T) {
-	p := DefaultParam()
-
-	t.Run("returns input when fuzz disabled", func(t *testing.T) {
-		p.EnableFuzz = false
-		got := p.ApplyFuzz(5.0, 0, false)
-		if got != 5.0 {
-			t.Errorf("expected 5.0, got=%v", got)
-		}
-	})
-
-	t.Run("returns input when interval below 2.5", func(t *testing.T) {
-		got := p.ApplyFuzz(2.3, 0, true)
-		if got != 2.3 {
-			t.Errorf("expected 2.3, got=%v", got)
-		}
-	})
-
-	t.Run("fuzz result within expected range at boundary 2.5", func(t *testing.T) {
-		p.EnableFuzz = true
-		p.seed = "test-seed-2.5"
-		minInterval, maxInterval := getFuzzRange(3, 0, p.MaximumInterval)
-		got := p.ApplyFuzz(2.5, 0, true)
-		if got < float64(minInterval) || got > float64(maxInterval) {
-			t.Errorf("expected result in [%d, %d], got=%v", minInterval, maxInterval, got)
-		}
-	})
-
-	t.Run("fuzz result within expected range for interval 3", func(t *testing.T) {
-		p.EnableFuzz = true
-		p.seed = "test-seed-3"
-		minInterval, maxInterval := getFuzzRange(3, 0, p.MaximumInterval)
-		got := p.ApplyFuzz(3.0, 0, true)
-		if got < float64(minInterval) || got > float64(maxInterval) {
-			t.Errorf("expected result in [%d, %d], got=%v", minInterval, maxInterval, got)
-		}
-	})
-
-	t.Run("fuzz result clamped by maximum interval", func(t *testing.T) {
-		p.EnableFuzz = true
-		p.MaximumInterval = 5
-		p.seed = "test-seed-max"
-		got := p.ApplyFuzz(100.0, 0, true)
-		if got > 5 {
-			t.Errorf("expected result <= 5 (MaximumInterval), got=%v", got)
-		}
-	})
-
-	t.Run("fuzz result respects elapsed days floor", func(t *testing.T) {
-		p.EnableFuzz = true
-		p.MaximumInterval = 36500
-		p.seed = "test-seed-elapsed"
-		interval := 5.0
-		elapsedDays := 4.0
-		minInterval, maxInterval := getFuzzRange(interval, elapsedDays, p.MaximumInterval)
-		got := p.ApplyFuzz(interval, elapsedDays, true)
-		if got < float64(minInterval) || got > float64(maxInterval) {
-			t.Errorf("expected result in [%d, %d], got=%v", minInterval, maxInterval, got)
-		}
-	})
-
-	t.Run("fuzz is deterministic for same seed", func(t *testing.T) {
-		p.EnableFuzz = true
-		p.seed = "deterministic-seed"
-		first := p.ApplyFuzz(10.0, 0, true)
-		p.seed = "deterministic-seed"
-		second := p.ApplyFuzz(10.0, 0, true)
-		if first != second {
-			t.Errorf("expected deterministic results: first=%v, second=%v", first, second)
-		}
-	})
-
-	t.Run("fuzz range at large interval", func(t *testing.T) {
-		p.EnableFuzz = true
-		p.MaximumInterval = 36500
-		p.seed = "test-seed-large"
-		minInterval, maxInterval := getFuzzRange(100.0, 0, p.MaximumInterval)
-		got := p.ApplyFuzz(100.0, 0, true)
-		if got < float64(minInterval) || got > float64(maxInterval) {
-			t.Errorf("expected result in [%d, %d], got=%v", minInterval, maxInterval, got)
-		}
-		if maxInterval > 36500 {
-			t.Errorf("expected maxInterval <= 36500, got=%d", maxInterval)
-		}
-	})
-}
-
-func TestErrorType(t *testing.T) {
-	t.Run("Error() returns message", func(t *testing.T) {
-		err := &Error{Code: ErrCodeInvalidInput, Message: "test message"}
-		if err.Error() != "test message" {
-			t.Errorf("expected 'test message', got=%v", err.Error())
-		}
-	})
-
-	t.Run("errors.Is matches by code", func(t *testing.T) {
-		err := &Error{Code: ErrCodeManualRating, Message: "different instance"}
-		if !errors.Is(err, ErrManualRating) {
-			t.Error("expected errors.Is to match by code")
-		}
-	})
-
-	t.Run("errors.Is does not match different code", func(t *testing.T) {
-		err := &Error{Code: ErrCodeInvalidRating, Message: "different code"}
-		if errors.Is(err, ErrManualRating) {
-			t.Error("expected errors.Is NOT to match different code")
-		}
-	})
-
-	t.Run("errors.Is works with singleton vars", func(t *testing.T) {
-		if !errors.Is(ErrManualRating, ErrManualRating) {
-			t.Error("expected singleton to match itself")
-		}
-		if !errors.Is(ErrInvalidRating, ErrInvalidRating) {
-			t.Error("expected singleton to match itself")
-		}
-		if errors.Is(ErrManualRating, ErrInvalidRating) {
-			t.Error("expected different singletons NOT to match")
-		}
-	})
-
-	t.Run("errors.As extracts code", func(t *testing.T) {
-		err := &Error{Code: ErrCodeManualStateRequired, Message: "state needed"}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatal("expected errors.As to extract *Error")
-		}
-		if fsrsErr.Code != ErrCodeManualStateRequired {
-			t.Errorf("expected code %v, got=%v", ErrCodeManualStateRequired, fsrsErr.Code)
-		}
-	})
-
-	t.Run("existing error messages unchanged", func(t *testing.T) {
-		if ErrManualRating.Error() != "fsrs: cannot rollback a manual rating" {
-			t.Errorf("ErrManualRating message changed: %v", ErrManualRating.Error())
-		}
-		if ErrInvalidRating.Error() != "fsrs: invalid rating for rollback" {
-			t.Errorf("ErrInvalidRating message changed: %v", ErrInvalidRating.Error())
-		}
-		if ErrManualStateRequired.Error() != "fsrs: state is required for manual rating" {
-			t.Errorf("ErrManualStateRequired message changed: %v", ErrManualStateRequired.Error())
-		}
-		if ErrManualDueRequired.Error() != "fsrs: due is required for manual rating when state is not New" {
-			t.Errorf("ErrManualDueRequired message changed: %v", ErrManualDueRequired.Error())
-		}
-	})
-
-	t.Run("errors.Is returns false for non-*Error target", func(t *testing.T) {
-		if errors.Is(ErrManualRating, errors.New("fsrs: cannot rollback a manual rating")) {
-			t.Error("expected errors.Is to return false for plain error")
-		}
-	})
-
-	t.Run("Is handles nil safely", func(t *testing.T) {
-		var nilErr *Error
-		if errors.Is(ErrManualRating, nilErr) {
-			t.Error("expected errors.Is with nil target to return false")
-		}
-		if nilErr.Is(ErrManualRating) {
-			t.Error("expected nil receiver.Is to return false")
-		}
-		if nilErr.Error() != "fsrs: <nil>" {
-			t.Errorf("expected nil Error() to return 'fsrs: <nil>', got=%v", nilErr.Error())
-		}
-	})
-
-	t.Run("errors.As negative case", func(t *testing.T) {
-		plainErr := errors.New("plain error")
-		var fsrsErr *Error
-		if errors.As(plainErr, &fsrsErr) {
-			t.Error("expected errors.As to fail for non-*Error")
-		}
-	})
-
-	t.Run("sentinel codes match constants", func(t *testing.T) {
-		if ErrManualRating.Code != ErrCodeManualRating {
-			t.Errorf("ErrManualRating.Code=%v, want ErrCodeManualRating=%v", ErrManualRating.Code, ErrCodeManualRating)
-		}
-		if ErrInvalidRating.Code != ErrCodeInvalidRating {
-			t.Errorf("ErrInvalidRating.Code=%v, want ErrCodeInvalidRating=%v", ErrInvalidRating.Code, ErrCodeInvalidRating)
-		}
-		if ErrManualStateRequired.Code != ErrCodeManualStateRequired {
-			t.Errorf("ErrManualStateRequired.Code=%v, want ErrCodeManualStateRequired=%v", ErrManualStateRequired.Code, ErrCodeManualStateRequired)
-		}
-		if ErrManualDueRequired.Code != ErrCodeManualDueRequired {
-			t.Errorf("ErrManualDueRequired.Code=%v, want ErrCodeManualDueRequired=%v", ErrManualDueRequired.Code, ErrCodeManualDueRequired)
-		}
-	})
-}
-
-func TestNextInputValidation(t *testing.T) {
-	f := NewFSRS(DefaultParam())
-	card := NewCard(time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC))
-	now := time.Date(2024, 1, 2, 0, 0, 0, 0, time.UTC)
-
-	t.Run("Manual grade returns error", func(t *testing.T) {
-		_, err := f.Next(card, now, Manual)
-		if err == nil {
-			t.Fatal("expected error for Manual grade")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("out of range grade returns error", func(t *testing.T) {
-		_, err := f.Next(card, now, Rating(5))
-		if err == nil {
-			t.Fatal("expected error for grade 5")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("invalid card state returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = State(42)
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for invalid state")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("NaN stability for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = math.NaN()
-		badCard.Difficulty = 5.0
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for NaN stability")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("zero stability for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = 0
-		badCard.Difficulty = 5.0
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for zero stability on Review card")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("Inf difficulty for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Learning
-		badCard.Stability = 1.0
-		badCard.Difficulty = math.Inf(1)
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for Inf difficulty")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("difficulty below dMin for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = 5.0
-		badCard.Difficulty = 0.5
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for difficulty below dMin")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("LastReview after now returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = 10.0
-		badCard.Difficulty = 5.0
-		badCard.LastReview = time.Date(2024, 1, 3, 0, 0, 0, 0, time.UTC)
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for future LastReview")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("valid New card succeeds", func(t *testing.T) {
-		_, err := f.Next(card, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	})
-
-	t.Run("valid Review card succeeds", func(t *testing.T) {
-		reviewCard := card
-		reviewCard.State = Review
-		reviewCard.Stability = 10.0
-		reviewCard.Difficulty = 5.0
-		reviewCard.LastReview = time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		_, err := f.Next(reviewCard, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	})
-
-	t.Run("valid Learning card succeeds", func(t *testing.T) {
-		learningCard := card
-		learningCard.State = Learning
-		learningCard.Stability = 2.0
-		learningCard.Difficulty = 5.0
-		learningCard.LastReview = time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		_, err := f.Next(learningCard, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	})
-
-	t.Run("valid Relearning card succeeds", func(t *testing.T) {
-		relearningCard := card
-		relearningCard.State = Relearning
-		relearningCard.Stability = 2.0
-		relearningCard.Difficulty = 5.0
-		relearningCard.LastReview = time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		_, err := f.Next(relearningCard, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	})
-
-	t.Run("negative stability for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = -5.0
-		badCard.Difficulty = 5.0
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for negative stability")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("stability in (0, sMin) for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = 0.0005
-		badCard.Difficulty = 5.0
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for stability below sMin")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("negative difficulty for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = 5.0
-		badCard.Difficulty = -1.0
-		_, err := f.Next(badCard, now, Good)
-		if err == nil {
-			t.Fatal("expected error for negative difficulty")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("stability at sMin boundary succeeds", func(t *testing.T) {
-		boundaryCard := card
-		boundaryCard.State = Review
-		boundaryCard.Stability = sMin
-		boundaryCard.Difficulty = 5.0
-		boundaryCard.LastReview = time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		_, err := f.Next(boundaryCard, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error at sMin boundary: %v", err)
-		}
-	})
-
-	t.Run("difficulty at dMin boundary succeeds", func(t *testing.T) {
-		boundaryCard := card
-		boundaryCard.State = Review
-		boundaryCard.Stability = 5.0
-		boundaryCard.Difficulty = dMin
-		boundaryCard.LastReview = time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		_, err := f.Next(boundaryCard, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error at dMin boundary: %v", err)
-		}
-	})
-
-	t.Run("LastReview equal to now succeeds", func(t *testing.T) {
-		eqCard := card
-		eqCard.State = Review
-		eqCard.Stability = 10.0
-		eqCard.Difficulty = 5.0
-		eqCard.LastReview = now
-		_, err := f.Next(eqCard, now, Good)
-		if err != nil {
-			t.Fatalf("unexpected error when LastReview == now: %v", err)
-		}
-	})
-
-	t.Run("Again grade boundary succeeds", func(t *testing.T) {
-		_, err := f.Next(card, now, Again)
-		if err != nil {
-			t.Fatalf("unexpected error for Again: %v", err)
-		}
-	})
-
-	t.Run("Easy grade boundary succeeds", func(t *testing.T) {
-		_, err := f.Next(card, now, Easy)
-		if err != nil {
-			t.Fatalf("unexpected error for Easy: %v", err)
-		}
-	})
-}
-
-func TestRepeatInputValidation(t *testing.T) {
-	f := NewFSRS(DefaultParam())
-	card := NewCard(time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC))
-	now := time.Date(2024, 1, 2, 0, 0, 0, 0, time.UTC)
-
-	t.Run("invalid card state returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = State(99)
-		_, err := f.Repeat(badCard, now)
-		if err == nil {
-			t.Fatal("expected error for invalid state")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("valid New card succeeds", func(t *testing.T) {
-		_, err := f.Repeat(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	})
-
-	t.Run("valid Review card succeeds", func(t *testing.T) {
-		reviewCard := card
-		reviewCard.State = Review
-		reviewCard.Stability = 10.0
-		reviewCard.Difficulty = 5.0
-		reviewCard.LastReview = time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		_, err := f.Repeat(reviewCard, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	})
-
-	t.Run("valid Learning card succeeds", func(t *testing.T) {
-		learningCard := card
-		learningCard.State = Learning
-		learningCard.Stability = 2.0
-		learningCard.Difficulty = 5.0
-		learningCard.LastReview = time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		_, err := f.Repeat(learningCard, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	})
-
-	t.Run("valid Relearning card succeeds", func(t *testing.T) {
-		relearningCard := card
-		relearningCard.State = Relearning
-		relearningCard.Stability = 2.0
-		relearningCard.Difficulty = 5.0
-		relearningCard.LastReview = time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)
-		_, err := f.Repeat(relearningCard, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	})
-
-	t.Run("NaN stability for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = math.NaN()
-		badCard.Difficulty = 5.0
-		_, err := f.Repeat(badCard, now)
-		if err == nil {
-			t.Fatal("expected error for NaN stability")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("zero stability for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = 0
-		badCard.Difficulty = 5.0
-		_, err := f.Repeat(badCard, now)
-		if err == nil {
-			t.Fatal("expected error for zero stability")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("difficulty below dMin for non-New card returns error", func(t *testing.T) {
-		badCard := card
-		badCard.State = Review
-		badCard.Stability = 5.0
-		badCard.Difficulty = 0.5
-		_, err := f.Repeat(badCard, now)
-		if err == nil {
-			t.Fatal("expected error for difficulty below dMin")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-}
-
-func TestRetrievabilityInputValidation(t *testing.T) {
-	f := NewFSRS(DefaultParam())
-	now := time.Date(2024, 1, 2, 0, 0, 0, 0, time.UTC)
-
-	t.Run("New card returns 0 without error", func(t *testing.T) {
-		card := NewCard(time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC))
-		r, err := f.Retrievability(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if r != 0 {
-			t.Errorf("expected 0, got=%v", r)
-		}
-	})
-
-	t.Run("card with zero LastReview returns 0 without error", func(t *testing.T) {
-		card := Card{State: Review, Stability: 10.0, Difficulty: 5.0}
-		r, err := f.Retrievability(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if r != 0 {
-			t.Errorf("expected 0, got=%v", r)
-		}
-	})
-
-	t.Run("invalid state returns error", func(t *testing.T) {
-		card := Card{State: State(99), Stability: 10.0, Difficulty: 5.0, LastReview: now.Add(-24 * time.Hour)}
-		_, err := f.Retrievability(card, now)
-		if err == nil {
-			t.Fatal("expected error for invalid state")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("NaN stability returns error", func(t *testing.T) {
-		card := Card{State: Review, Stability: math.NaN(), Difficulty: 5.0, LastReview: now.Add(-24 * time.Hour)}
-		_, err := f.Retrievability(card, now)
-		if err == nil {
-			t.Fatal("expected error for NaN stability")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("negative stability returns error", func(t *testing.T) {
-		card := Card{State: Review, Stability: -1.0, Difficulty: 5.0, LastReview: now.Add(-24 * time.Hour)}
-		_, err := f.Retrievability(card, now)
-		if err == nil {
-			t.Fatal("expected error for negative stability")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("valid Review card succeeds", func(t *testing.T) {
-		card := Card{State: Review, Stability: 10.0, Difficulty: 5.0, LastReview: now.Add(-24 * time.Hour)}
-		r, err := f.Retrievability(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if r <= 0 || r > 1 {
-			t.Errorf("expected retrievability in (0,1], got=%v", r)
-		}
-	})
-
-	t.Run("zero stability returns error", func(t *testing.T) {
-		card := Card{State: Review, Stability: 0, Difficulty: 5.0, LastReview: now.Add(-24 * time.Hour)}
-		_, err := f.Retrievability(card, now)
-		if err == nil {
-			t.Fatal("expected error for zero stability")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("valid Learning card succeeds", func(t *testing.T) {
-		card := Card{State: Learning, Stability: 2.0, Difficulty: 5.0, LastReview: now.Add(-24 * time.Hour)}
-		r, err := f.Retrievability(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if r <= 0 || r > 1 {
-			t.Errorf("expected retrievability in (0,1], got=%v", r)
-		}
-	})
-
-	t.Run("valid Relearning card succeeds", func(t *testing.T) {
-		card := Card{State: Relearning, Stability: 2.0, Difficulty: 5.0, LastReview: now.Add(-24 * time.Hour)}
-		r, err := f.Retrievability(card, now)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if r <= 0 || r > 1 {
-			t.Errorf("expected retrievability in (0,1], got=%v", r)
-		}
-	})
-}
-
-func TestValidateResult(t *testing.T) {
-	t.Run("valid card passes", func(t *testing.T) {
-		card := Card{Stability: 5.0, Difficulty: 5.0}
-		if err := validateResult(card); err != nil {
-			t.Errorf("expected nil, got=%v", err)
-		}
-	})
-
-	t.Run("stability below sMin returns error", func(t *testing.T) {
-		card := Card{Stability: 0.0001, Difficulty: 5.0}
-		err := validateResult(card)
-		if err == nil {
-			t.Fatal("expected error for stability below sMin")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("NaN stability returns error", func(t *testing.T) {
-		card := Card{Stability: math.NaN(), Difficulty: 5.0}
-		err := validateResult(card)
-		if err == nil {
-			t.Fatal("expected error for NaN stability")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("difficulty below dMin returns error", func(t *testing.T) {
-		card := Card{Stability: 5.0, Difficulty: 0.5}
-		err := validateResult(card)
-		if err == nil {
-			t.Fatal("expected error for difficulty below dMin")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("NaN difficulty returns error", func(t *testing.T) {
-		card := Card{Stability: 5.0, Difficulty: math.NaN()}
-		err := validateResult(card)
-		if err == nil {
-			t.Fatal("expected error for NaN difficulty")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("stability at sMin boundary passes", func(t *testing.T) {
-		card := Card{Stability: sMin, Difficulty: 5.0}
-		if err := validateResult(card); err != nil {
-			t.Errorf("expected nil at boundary, got=%v", err)
-		}
-	})
-
-	t.Run("difficulty at dMin boundary passes", func(t *testing.T) {
-		card := Card{Stability: 5.0, Difficulty: dMin}
-		if err := validateResult(card); err != nil {
-			t.Errorf("expected nil at boundary, got=%v", err)
-		}
-	})
-}
-
-func TestValidateCardDirect(t *testing.T) {
-	now := time.Date(2024, 1, 2, 0, 0, 0, 0, time.UTC)
-
-	t.Run("valid New card", func(t *testing.T) {
-		card := Card{State: New}
-		if err := validateCard(card, now); err != nil {
-			t.Errorf("expected nil, got=%v", err)
-		}
-	})
-
-	t.Run("valid Review card", func(t *testing.T) {
-		card := Card{State: Review, Stability: 5.0, Difficulty: 5.0}
-		if err := validateCard(card, now); err != nil {
-			t.Errorf("expected nil, got=%v", err)
-		}
-	})
-
-	t.Run("New card skips stability check", func(t *testing.T) {
-		card := Card{State: New, Stability: 0, Difficulty: 0}
-		if err := validateCard(card, now); err != nil {
-			t.Errorf("expected nil for New card with zero fields, got=%v", err)
-		}
-	})
-
-	t.Run("invalid state", func(t *testing.T) {
-		card := Card{State: State(42)}
-		err := validateCard(card, now)
-		if err == nil {
-			t.Fatal("expected error for invalid state")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) {
-			t.Fatalf("expected *Error, got %T", err)
-		}
-		if fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", fsrsErr.Code)
-		}
-	})
-
-	t.Run("NaN stability non-New", func(t *testing.T) {
-		card := Card{State: Review, Stability: math.NaN(), Difficulty: 5.0}
-		err := validateCard(card, now)
-		if err == nil {
-			t.Fatal("expected error")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", err)
-		}
-	})
-
-	t.Run("stability below sMin non-New", func(t *testing.T) {
-		card := Card{State: Review, Stability: 0.0005, Difficulty: 5.0}
-		err := validateCard(card, now)
-		if err == nil {
-			t.Fatal("expected error")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", err)
-		}
-	})
-
-	t.Run("stability at sMin non-New passes", func(t *testing.T) {
-		card := Card{State: Review, Stability: sMin, Difficulty: 5.0}
-		if err := validateCard(card, now); err != nil {
-			t.Errorf("expected nil at sMin boundary, got=%v", err)
-		}
-	})
-
-	t.Run("NaN difficulty non-New", func(t *testing.T) {
-		card := Card{State: Review, Stability: 5.0, Difficulty: math.NaN()}
-		err := validateCard(card, now)
-		if err == nil {
-			t.Fatal("expected error")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", err)
-		}
-	})
-
-	t.Run("difficulty below dMin non-New", func(t *testing.T) {
-		card := Card{State: Review, Stability: 5.0, Difficulty: 0.5}
-		err := validateCard(card, now)
-		if err == nil {
-			t.Fatal("expected error")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", err)
-		}
-	})
-
-	t.Run("difficulty at dMin non-New passes", func(t *testing.T) {
-		card := Card{State: Review, Stability: 5.0, Difficulty: dMin}
-		if err := validateCard(card, now); err != nil {
-			t.Errorf("expected nil at dMin boundary, got=%v", err)
-		}
-	})
-
-	t.Run("LastReview after now", func(t *testing.T) {
-		card := Card{State: Review, Stability: 5.0, Difficulty: 5.0, LastReview: now.Add(1 * time.Hour)}
-		err := validateCard(card, now)
-		if err == nil {
-			t.Fatal("expected error for future LastReview")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", err)
-		}
-	})
-
-	t.Run("LastReview equal to now passes", func(t *testing.T) {
-		card := Card{State: Review, Stability: 5.0, Difficulty: 5.0, LastReview: now}
-		if err := validateCard(card, now); err != nil {
-			t.Errorf("expected nil when LastReview == now, got=%v", err)
-		}
-	})
-
-	t.Run("zero LastReview passes", func(t *testing.T) {
-		card := Card{State: Review, Stability: 5.0, Difficulty: 5.0}
-		if err := validateCard(card, now); err != nil {
-			t.Errorf("expected nil for zero LastReview, got=%v", err)
-		}
-	})
-}
-
-func TestValidateRatingDirect(t *testing.T) {
-	t.Run("Again passes", func(t *testing.T) {
-		if err := validateRating(Again); err != nil {
-			t.Errorf("expected nil for Again, got=%v", err)
-		}
-	})
-
-	t.Run("Easy passes", func(t *testing.T) {
-		if err := validateRating(Easy); err != nil {
-			t.Errorf("expected nil for Easy, got=%v", err)
-		}
-	})
-
-	t.Run("Good passes", func(t *testing.T) {
-		if err := validateRating(Good); err != nil {
-			t.Errorf("expected nil for Good, got=%v", err)
-		}
-	})
-
-	t.Run("Hard passes", func(t *testing.T) {
-		if err := validateRating(Hard); err != nil {
-			t.Errorf("expected nil for Hard, got=%v", err)
-		}
-	})
-
-	t.Run("Manual returns error", func(t *testing.T) {
-		err := validateRating(Manual)
-		if err == nil {
-			t.Fatal("expected error for Manual")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", err)
-		}
-	})
-
-	t.Run("Rating(5) returns error", func(t *testing.T) {
-		err := validateRating(Rating(5))
-		if err == nil {
-			t.Fatal("expected error for Rating(5)")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", err)
-		}
-	})
-
-	t.Run("Rating(0) returns error", func(t *testing.T) {
-		err := validateRating(Rating(0))
-		if err == nil {
-			t.Fatal("expected error for Rating(0)")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidInput {
-			t.Errorf("expected ErrCodeInvalidInput, got=%v", err)
-		}
-	})
-}
-
-func TestGetFuzzRange(t *testing.T) {
-	t.Run("small interval uses minimal delta", func(t *testing.T) {
-		minInterval, maxInterval := getFuzzRange(3, 0, 36500)
-		if minInterval > maxInterval {
-			t.Errorf("minInterval > maxInterval: %d > %d", minInterval, maxInterval)
-		}
-		if minInterval < 2 {
-			t.Errorf("expected minInterval >= 2, got=%d", minInterval)
-		}
-	})
-
-	t.Run("clamped by maximum interval", func(t *testing.T) {
-		_, maxInterval := getFuzzRange(100000, 0, 10)
-		if maxInterval > 10 {
-			t.Errorf("expected maxInterval <= 10, got=%d", maxInterval)
-		}
-	})
-
-	t.Run("elapsed days floor when interval exceeds elapsed", func(t *testing.T) {
-		minInterval, _ := getFuzzRange(5, 4, 36500)
-		if minInterval < 5 {
-			t.Errorf("expected minInterval >= 5 (elapsedDays+1), got=%d", minInterval)
-		}
-	})
-
-	t.Run("min never exceeds max", func(t *testing.T) {
-		for _, tc := range []struct {
-		interval    float64
-		elapsed float64
-		maxInterval float64
-		}{
-			{2.5, 0, 36500},
-			{7.0, 0, 36500},
-			{20.0, 0, 36500},
-			{5.0, 10, 36500},
-			{3.0, 0, 2},
-		} {
-			minInterval, maxInterval := getFuzzRange(tc.interval, tc.elapsed, tc.maxInterval)
-			if minInterval > maxInterval {
-				t.Errorf("interval=%v elapsed=%v maxInterval=%v: minInterval=%d > maxInterval=%d", tc.interval, tc.elapsed, tc.maxInterval, minInterval, maxInterval)
-			}
-		}
-	})
-}
-
-func TestFuzzRanges(t *testing.T) {
-	ranges := FuzzRanges()
-
-	if len(ranges) != 3 {
-		t.Fatalf("expected 3 fuzz ranges, got %d", len(ranges))
-	}
-
-	expected := []fuzzRange{
-		{Start: 2.5, End: 7.0, Factor: 0.15},
-		{Start: 7.0, End: 20.0, Factor: 0.10},
-		{Start: 20.0, End: math.Inf(1), Factor: 0.05},
-	}
-
-	for i, r := range ranges {
-		if r.Start != expected[i].Start || r.End != expected[i].End || r.Factor != expected[i].Factor {
-			t.Errorf("range[%d] got={%.2f,%.2f,%.2f} want={%.2f,%.2f,%.2f}",
-				i, r.Start, r.End, r.Factor, expected[i].Start, expected[i].End, expected[i].Factor)
-		}
-	}
-
-	t.Run("defensive copy", func(t *testing.T) {
-		ranges2 := FuzzRanges()
-		ranges[0].Start = 999
-		if ranges2[0].Start == 999 {
-			t.Error("mutating returned slice should not affect internal state")
-		}
-	})
-}
-
-func TestConvertV45WeightsParity(t *testing.T) {
-	v45 := [17]float64{
-		0.4, 0.6, 2.4, 5.8, 4.93, 0.94, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61,
-	}
-	expected := [21]float64{
-		0.4, 0.6, 2.4, 5.8, 6.81, 0.44675013, 1.36, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61, 0.0, 0.0, 0.0, 0.5,
-	}
-
-	w, err := ConvertV45Weights(v45)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	const tolerance = 1e-6
-	for i := 0; i < 21; i++ {
-		if math.Abs(w[i]-expected[i]) > tolerance {
-			t.Errorf("W[%d]: expected ≈ %v, got %v", i, expected[i], w[i])
-		}
-	}
-}
-
-func TestDefaultWeights(t *testing.T) {
-	w := DefaultWeights()
-	if len(w) != 21 {
-		t.Fatalf("expected 21 weights, got %d", len(w))
-	}
-	for i, val := range w {
-		if math.IsNaN(val) || math.IsInf(val, 0) {
-			t.Errorf("W[%d]: expected finite, got %v", i, val)
-		}
-	}
-}
-
-func TestConvertV45WeightsInvalidInput(t *testing.T) {
-	t.Run("NaN input", func(t *testing.T) {
-		v45 := [17]float64{
-			0.4, 0.6, 2.4, 5.8, 4.93, math.NaN(), 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61,
-		}
-		_, err := ConvertV45Weights(v45)
-		if err == nil {
-			t.Fatal("expected error, got nil")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidWeightsValue {
-			t.Errorf("expected ErrCodeInvalidWeightsValue, got=%v", err)
-		}
-		if !errors.Is(err, ErrInvalidWeightsValue) {
-			t.Errorf("expected errors.Is(err, ErrInvalidWeightsValue), got=%v", err)
-		}
-	})
-
-	t.Run("Inf input", func(t *testing.T) {
-		v45 := [17]float64{
-			0.4, 0.6, 2.4, 5.8, 4.93, 0.94, math.Inf(1), 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61,
-		}
-		_, err := ConvertV45Weights(v45)
-		if err == nil {
-			t.Fatal("expected error, got nil")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidWeightsValue {
-			t.Errorf("expected ErrCodeInvalidWeightsValue, got=%v", err)
-		}
-		if !errors.Is(err, ErrInvalidWeightsValue) {
-			t.Errorf("expected errors.Is(err, ErrInvalidWeightsValue), got=%v", err)
-		}
-	})
-
-	t.Run("finite input producing NaN in conversion", func(t *testing.T) {
-		v45 := [17]float64{
-			0.4, 0.6, 2.4, 5.8, 4.93, -1000.0, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61,
-		}
-		_, err := ConvertV45Weights(v45)
-		if err == nil {
-			t.Fatal("expected error for conversion-produced NaN, got nil")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidWeightsValue {
-			t.Errorf("expected ErrCodeInvalidWeightsValue, got=%v", err)
-		}
-		if !errors.Is(err, ErrInvalidWeightsValue) {
-			t.Errorf("expected errors.Is(err, ErrInvalidWeightsValue), got=%v", err)
-		}
-	})
-
-	t.Run("all-zero input", func(t *testing.T) {
-		v45 := [17]float64{}
-		w, err := ConvertV45Weights(v45)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if w[4] != 0.0 {
-			t.Errorf("expected W[4] = 0.0, got %v", w[4])
-		}
-		if w[5] != 0.0 {
-			t.Errorf("expected W[5] = 0.0, got %v", w[5])
-		}
-		if w[6] != 0.5 {
-			t.Errorf("expected W[6] = 0.5, got %v", w[6])
-		}
-		if w[20] != 0.5 {
-			t.Errorf("expected W[20] = 0.5, got %v", w[20])
-		}
-	})
-}
-
-func TestConvertV45WeightsIndices(t *testing.T) {
-	v45 := [17]float64{
-		0.4, 0.6, 2.4, 5.8, 4.93, 0.94, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61,
-	}
-
-	w, err := ConvertV45Weights(v45)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	// W[0..3] unchanged
-	for i := 0; i < 4; i++ {
-		if w[i] != v45[i] {
-			t.Errorf("W[%d]: expected unchanged %v, got %v", i, v45[i], w[i])
-		}
-	}
-
-	// W[4] = W[5] * 2.0 + W[4]  (order: reads old W[5] before it is overwritten)
-	expectedW4 := v45[5]*2.0 + v45[4]
-	if math.Abs(w[4]-expectedW4) > 1e-6 {
-		t.Errorf("W[4]: expected ≈ %v, got %v", expectedW4, w[4])
-	}
-
-	// W[5] = ln(W[5] * 3.0 + 1.0) / 3.0
-	expectedW5 := math.Log(v45[5]*3.0+1.0) / 3.0
-	if math.Abs(w[5]-expectedW5) > 1e-6 {
-		t.Errorf("W[5]: expected ≈ %v, got %v", expectedW5, w[5])
-	}
-
-	// W[6] = W[6] + 0.5
-	expectedW6 := v45[6] + 0.5
-	if math.Abs(w[6]-expectedW6) > 1e-6 {
-		t.Errorf("W[6]: expected ≈ %v, got %v", expectedW6, w[6])
-	}
-
-	// W[7..16] unchanged
-	for i := 7; i < 17; i++ {
-		if w[i] != v45[i] {
-			t.Errorf("W[%d]: expected unchanged %v, got %v", i, v45[i], w[i])
-		}
-	}
-
-	// W[17..19] = 0.0
-	for i := 17; i < 20; i++ {
-		if w[i] != 0.0 {
-			t.Errorf("W[%d]: expected 0.0, got %v", i, w[i])
-		}
-	}
-
-	// W[20] = fsrs5DefaultDecay (0.5)
-	if w[20] != 0.5 {
-		t.Errorf("W[20]: expected 0.5, got %v", w[20])
-	}
-}
-
-func TestConvertV45WeightsProducesValidFSRS(t *testing.T) {
-	v45 := [17]float64{
-		0.4, 0.6, 2.4, 5.8, 4.93, 0.94, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61,
-	}
-
-	w, err := ConvertV45Weights(v45)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	// Verify converted weights produce a valid FSRS instance
-	p := DefaultParam()
-	p.W = w
-	if err := p.Validate(); err != nil {
-		t.Errorf("converted weights should pass Validate: %v", err)
-	}
-
-	f := NewFSRS(p)
-	card := NewCard()
-	now := time.Now()
-
-	// Verify Repeat works with converted weights
-	result, err := f.Repeat(card, now)
-	if err != nil {
-		t.Errorf("Repeat should work with converted weights: %v", err)
-	}
-	if len(result) != 4 {
-		t.Errorf("expected 4 scheduling cards, got %d", len(result))
-	}
-
-	// Verify Retrievability works with converted weights
-	_, err = f.Retrievability(card, now)
-	if err != nil {
-		t.Errorf("Retrievability should work with converted weights: %v", err)
-	}
-}
-
-func TestMigrateWeights(t *testing.T) {
-	t.Run("length 17 parity", func(t *testing.T) {
-		v17 := []float64{
-			0.4, 0.6, 2.4, 5.8, 4.93, 0.94, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61,
-		}
-		w, err := MigrateWeights(v17)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(w) != 21 {
-			t.Fatalf("expected length 21, got %d", len(w))
-		}
-		if math.Abs(w[4]-6.81) > 1e-6 {
-			t.Errorf("expected W[4] ≈ 6.81, got %v", w[4])
-		}
-	})
-
-	t.Run("length 19 migration", func(t *testing.T) {
-		v19 := []float64{
-			0.4, 0.6, 2.4, 5.8, 4.93, 0.94, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61, 1.0, 2.0,
-		}
-		w, err := MigrateWeights(v19)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if w[19] != 0.0 {
-			t.Errorf("expected W[19] = 0.0, got %v", w[19])
-		}
-		if w[20] != 0.5 {
-			t.Errorf("expected W[20] = 0.5, got %v", w[20])
-		}
-	})
-
-	t.Run("length 21 migration", func(t *testing.T) {
-		v21 := make([]float64, 21)
-		for i := 0; i < 21; i++ {
-			v21[i] = float64(i) + 0.1
-		}
-		w, err := MigrateWeights(v21)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		for i := 0; i < 21; i++ {
-			if w[i] != v21[i] {
-				t.Errorf("expected W[%d] = %v, got %v", i, v21[i], w[i])
-			}
-		}
-	})
-
-	t.Run("invalid length", func(t *testing.T) {
-		_, err := MigrateWeights([]float64{1.0, 2.0})
-		if err == nil {
-			t.Fatal("expected error, got nil")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidWeightsLength {
-			t.Errorf("expected ErrCodeInvalidWeightsLength, got=%v", err)
-		}
-		if !errors.Is(err, ErrInvalidWeightsLength) {
-			t.Errorf("expected errors.Is(err, ErrInvalidWeightsLength), got=%v", err)
-		}
-	})
-
-	t.Run("NaN in slice", func(t *testing.T) {
-		v19 := []float64{
-			0.4, 0.6, 2.4, 5.8, 4.93, 0.94, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61, 1.0, math.NaN(),
-		}
-		_, err := MigrateWeights(v19)
-		if err == nil {
-			t.Fatal("expected error, got nil")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidWeightsValue {
-			t.Errorf("expected ErrCodeInvalidWeightsValue, got=%v", err)
-		}
-		if !errors.Is(err, ErrInvalidWeightsValue) {
-			t.Errorf("expected errors.Is(err, ErrInvalidWeightsValue), got=%v", err)
-		}
-	})
-
-	t.Run("Inf in length 19 slice", func(t *testing.T) {
-		v19 := []float64{
-			0.4, 0.6, 2.4, 5.8, 4.93, 0.94, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61, 1.0, math.Inf(1),
-		}
-		_, err := MigrateWeights(v19)
-		if err == nil {
-			t.Fatal("expected error, got nil")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidWeightsValue {
-			t.Errorf("expected ErrCodeInvalidWeightsValue, got=%v", err)
-		}
-		if !errors.Is(err, ErrInvalidWeightsValue) {
-			t.Errorf("expected errors.Is(err, ErrInvalidWeightsValue), got=%v", err)
-		}
-	})
-
-	t.Run("NaN in length 21 slice", func(t *testing.T) {
-		v21 := make([]float64, 21)
-		for i := range v21 {
-			v21[i] = float64(i) + 0.1
-		}
-		v21[20] = math.NaN()
-		_, err := MigrateWeights(v21)
-		if err == nil {
-			t.Fatal("expected error, got nil")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidWeightsValue {
-			t.Errorf("expected ErrCodeInvalidWeightsValue, got=%v", err)
-		}
-		if !errors.Is(err, ErrInvalidWeightsValue) {
-			t.Errorf("expected errors.Is(err, ErrInvalidWeightsValue), got=%v", err)
-		}
-	})
-
-	t.Run("empty slice", func(t *testing.T) {
-		_, err := MigrateWeights([]float64{})
-		if err == nil {
-			t.Fatal("expected error, got nil")
-		}
-		var fsrsErr *Error
-		if !errors.As(err, &fsrsErr) || fsrsErr.Code != ErrCodeInvalidWeightsLength {
-			t.Errorf("expected ErrCodeInvalidWeightsLength, got=%v", err)
-		}
-		if !errors.Is(err, ErrInvalidWeightsLength) {
-			t.Errorf("expected errors.Is(err, ErrInvalidWeightsLength), got=%v", err)
-		}
-	})
-
-	t.Run("sentinels are distinguishable", func(t *testing.T) {
-		_, lengthErr := MigrateWeights([]float64{})
-		_, valueErr := MigrateWeights([]float64{
-			0.4, 0.6, 2.4, 5.8, 4.93, 0.94, 0.86, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, math.NaN(),
-		})
-
-		if errors.Is(lengthErr, ErrInvalidWeightsValue) {
-			t.Error("length error should NOT match ErrInvalidWeightsValue")
-		}
-		if errors.Is(valueErr, ErrInvalidWeightsLength) {
-			t.Error("value error should NOT match ErrInvalidWeightsLength")
-		}
-	})
-}
-
-func TestConvertV5Weights(t *testing.T) {
-	v5 := [19]float64{
-		0.4, 0.6, 2.4, 5.8, 6.81, 0.44675013, 1.36, 0.01, 1.49, 0.14, 0.94, 2.18, 0.05, 0.34, 1.26, 0.29, 2.61, 1.0, 2.0,
-	}
-	w := ConvertV5Weights(v5)
-	for i := 0; i < 19; i++ {
-		if w[i] != v5[i] {
-			t.Errorf("W[%d]: expected %v, got %v", i, v5[i], w[i])
-		}
-	}
-	if w[19] != 0.0 {
-		t.Errorf("W[19]: expected 0.0, got %v", w[19])
-	}
-	if w[20] != 0.5 {
-		t.Errorf("W[20]: expected 0.5, got %v", w[20])
-	}
-}
-
-func TestDefaultLearningSteps(t *testing.T) {
-	steps := DefaultLearningSteps()
-	if want := []float64{1, 10}; !reflect.DeepEqual(steps, want) {
-		t.Errorf("expected %v, got %v", want, steps)
-	}
-	steps[0] = 999
-	if fresh := DefaultLearningSteps(); fresh[0] == 999 {
-		t.Error("mutating returned slice affected future calls")
-	}
-}
-
-func TestDefaultRelearningSteps(t *testing.T) {
-	steps := DefaultRelearningSteps()
-	if want := []float64{10}; !reflect.DeepEqual(steps, want) {
-		t.Errorf("expected %v, got %v", want, steps)
-	}
-	steps[0] = 999
-	if fresh := DefaultRelearningSteps(); fresh[0] == 999 {
-		t.Error("mutating returned slice affected future calls")
 	}
 }
