+++ testhelpers_test.go
@@ -1,0 +1,8 @@
+package fsrs
+
+import "math"
+
+func roundFloat(val float64, precision uint) float64 {
+	ratio := math.Pow(10, float64(precision))
+	return math.Round(val*ratio) / ratio
+}
